Introduce StmtHandler
Statement processing moves out of NodeScopeResolver's processStmtNode
if-chain into per-statement StmtHandler classes, dispatched through
StmtHandlerRegistry (memoized by statement class, mirroring
ExprHandlerRegistry). Overriding @throws points are applied centrally by
the dispatcher, so handlers return their computed throw points plainly.
