+++ src/Analyser/NodeScopeResolver.php
@@ -2,24 +2,16 @@
 
 namespace PHPStan\Analyser;
 
-use ArrayAccess;
 use Closure;
-use IteratorAggregate;
 use PhpParser\Comment\Doc;
-use PhpParser\Modifiers;
 use PhpParser\Node;
-use PhpParser\Node\Arg;
 use PhpParser\Node\AttributeGroup;
 use PhpParser\Node\ComplexType;
 use PhpParser\Node\Expr;
-use PhpParser\Node\Expr\Array_;
 use PhpParser\Node\Expr\ArrayDimFetch;
 use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\AssignRef;
-use PhpParser\Node\Expr\BinaryOp;
-use PhpParser\Node\Expr\BinaryOp\BooleanOr;
 use PhpParser\Node\Expr\CallLike;
-use PhpParser\Node\Expr\ConstFetch;
 use PhpParser\Node\Expr\FuncCall;
 use PhpParser\Node\Expr\List_;
 use PhpParser\Node\Expr\MethodCall;
@@ -30,86 +22,41 @@ use PhpParser\Node\Expr\StaticPropertyFetch;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Identifier;
 use PhpParser\Node\Name;
-use PhpParser\Node\Stmt\Break_;
 use PhpParser\Node\Stmt\Class_;
-use PhpParser\Node\Stmt\Continue_;
-use PhpParser\Node\Stmt\Do_;
-use PhpParser\Node\Stmt\Echo_;
-use PhpParser\Node\Stmt\For_;
 use PhpParser\Node\Stmt\Foreach_;
 use PhpParser\Node\Stmt\Goto_;
-use PhpParser\Node\Stmt\If_;
-use PhpParser\Node\Stmt\InlineHTML;
 use PhpParser\Node\Stmt\Return_;
 use PhpParser\Node\Stmt\Static_;
-use PhpParser\Node\Stmt\Switch_;
-use PhpParser\Node\Stmt\TryCatch;
-use PhpParser\Node\Stmt\Unset_;
-use PhpParser\Node\Stmt\While_;
 use PhpParser\NodeFinder;
 use PhpParser\NodeTraverser;
 use PHPStan\Analyser\ExprHandler\AssignHandler;
 use PHPStan\Analyser\ExprHandler\Helper\ClosureTypeResolver;
-use PHPStan\Analyser\ExprHandler\Helper\ImplicitToStringCallHelper;
-use PHPStan\Analyser\ExprHandler\Helper\MethodThrowPointHelper;
-use PHPStan\BetterReflection\Reflection\Adapter\ReflectionClass;
-use PHPStan\BetterReflection\Reflection\ReflectionEnum;
-use PHPStan\BetterReflection\Reflector\Reflector;
-use PHPStan\BetterReflection\SourceLocator\Ast\Strategy\NodeToReflection;
-use PHPStan\BetterReflection\SourceLocator\Located\LocatedSource;
 use PHPStan\DependencyInjection\AutowiredExtensions;
 use PHPStan\DependencyInjection\AutowiredParameter;
 use PHPStan\DependencyInjection\AutowiredService;
 use PHPStan\DependencyInjection\Container;
 use PHPStan\DependencyInjection\ExtensionsCollection;
 use PHPStan\File\FileHelper;
-use PHPStan\File\FileReader;
-use PHPStan\Node\BreaklessWhileLoopNode;
-use PHPStan\Node\CatchWithUnthrownExceptionNode;
-use PHPStan\Node\ClassConstantsNode;
-use PHPStan\Node\ClassMethodsNode;
-use PHPStan\Node\ClassPropertiesNode;
-use PHPStan\Node\ClassPropertyNode;
-use PHPStan\Node\ClassStatementsGatherer;
 use PHPStan\Node\ClosureReturnStatementsNode;
-use PHPStan\Node\DoWhileLoopConditionNode;
 use PHPStan\Node\ExecutionEndNode;
-use PHPStan\Node\Expr\ExistingArrayDimFetch;
-use PHPStan\Node\Expr\ForeachValueByRefExpr;
-use PHPStan\Node\Expr\NativeTypeExpr;
-use PHPStan\Node\Expr\OriginalForeachKeyExpr;
-use PHPStan\Node\Expr\OriginalForeachValueExpr;
-use PHPStan\Node\Expr\PropertyInitializationExpr;
 use PHPStan\Node\Expr\TypeExpr;
-use PHPStan\Node\Expr\UnsetOffsetExpr;
-use PHPStan\Node\FinallyExitPointsNode;
 use PHPStan\Node\FunctionCallableNode;
 use PHPStan\Node\FunctionCallExpressionNode;
-use PHPStan\Node\FunctionReturnStatementsNode;
 use PHPStan\Node\InArrowFunctionNode;
-use PHPStan\Node\InClassMethodNode;
-use PHPStan\Node\InClassNode;
 use PHPStan\Node\InClosureNode;
-use PHPStan\Node\InForeachNode;
-use PHPStan\Node\InFunctionNode;
 use PHPStan\Node\InPropertyHookNode;
 use PHPStan\Node\InstantiationCallableNode;
-use PHPStan\Node\InTraitNode;
 use PHPStan\Node\InvalidateExprNode;
 use PHPStan\Node\MethodCallableNode;
 use PHPStan\Node\MethodCallExpressionNode;
 use PHPStan\Node\MethodReturnStatementsNode;
-use PHPStan\Node\NoopExpressionNode;
 use PHPStan\Node\PropertyAssignNode;
 use PHPStan\Node\PropertyHookReturnStatementsNode;
 use PHPStan\Node\PropertyHookStatementNode;
 use PHPStan\Node\ReturnStatement;
 use PHPStan\Node\StaticMethodCallableNode;
 use PHPStan\Node\StaticMethodCallExpressionNode;
-use PHPStan\Node\SwitchConditionArm;
-use PHPStan\Node\SwitchConditionNode;
 use PHPStan\Node\UnreachableStatementNode;
-use PHPStan\Node\VariableAssignNode;
 use PHPStan\Node\VarTagChangedExpressionTypeNode;
 use PHPStan\Parser\ArrowFunctionArgVisitor;
 use PHPStan\Parser\ClosureArgVisitor;
@@ -124,7 +71,6 @@ use PHPStan\Reflection\Assertions;
 use PHPStan\Reflection\Callables\SimpleImpurePoint;
 use PHPStan\Reflection\Callables\SimpleThrowPoint;
 use PHPStan\Reflection\ClassReflection;
-use PHPStan\Reflection\ClassReflectionFactory;
 use PHPStan\Reflection\ExtendedMethodReflection;
 use PHPStan\Reflection\ExtendedParameterReflection;
 use PHPStan\Reflection\FunctionReflection;
@@ -136,7 +82,6 @@ use PHPStan\Reflection\Native\NativeParameterReflection;
 use PHPStan\Reflection\ParameterReflection;
 use PHPStan\Reflection\ParametersAcceptor;
 use PHPStan\Reflection\ParametersAcceptorSelector;
-use PHPStan\Reflection\Php\PhpFunctionFromParserNodeReflection;
 use PHPStan\Reflection\Php\PhpMethodFromParserNodeReflection;
 use PHPStan\Reflection\Php\PhpMethodReflection;
 use PHPStan\Reflection\Php\PhpPropertyReflection;
@@ -145,8 +90,6 @@ use PHPStan\Rules\Properties\ReadWritePropertiesExtension;
 use PHPStan\ShouldNotHappenException;
 use PHPStan\TrinaryLogic;
 use PHPStan\Type\ClosureType;
-use PHPStan\Type\Constant\ConstantIntegerType;
-use PHPStan\Type\Constant\ConstantStringType;
 use PHPStan\Type\FileTypeMapper;
 use PHPStan\Type\FunctionParameterClosureThisExtension;
 use PHPStan\Type\FunctionParameterClosureTypeExtension;
@@ -159,23 +102,18 @@ use PHPStan\Type\MethodParameterOutTypeExtension;
 use PHPStan\Type\MixedType;
 use PHPStan\Type\NeverType;
 use PHPStan\Type\NullType;
-use PHPStan\Type\ObjectType;
 use PHPStan\Type\ObjectWithoutClassType;
-use PHPStan\Type\ParserNodeTypeToPHPStanType;
 use PHPStan\Type\ResourceType;
 use PHPStan\Type\StaticMethodParameterClosureThisExtension;
 use PHPStan\Type\StaticMethodParameterClosureTypeExtension;
 use PHPStan\Type\StaticMethodParameterOutTypeExtension;
 use PHPStan\Type\StaticType;
-use PHPStan\Type\StaticTypeFactory;
 use PHPStan\Type\ThisType;
 use PHPStan\Type\Type;
 use PHPStan\Type\TypeCombinator;
 use PHPStan\Type\TypeTraverser;
 use PHPStan\Type\TypeUtils;
 use PHPStan\Type\UnionType;
-use Throwable;
-use Traversable;
 use function array_fill_keys;
 use function array_filter;
 use function array_key_exists;
@@ -195,16 +133,13 @@ use function sprintf;
 use function strtolower;
 use function trim;
 use function usort;
-use const PHP_VERSION_ID;
 
 #[AutowiredService]
 class NodeScopeResolver
 {
 
-	private const LOOP_SCOPE_ITERATIONS = 3;
-	private const GENERALIZE_AFTER_ITERATION = 1;
-	private const FOREACH_UNROLL_LIMIT = 16;
-	private const FOREACH_UNROLL_NESTED_LIMIT = 8;
+	public const LOOP_SCOPE_ITERATIONS = 3;
+	public const GENERALIZE_AFTER_ITERATION = 1;
 
 	/** @var array<string, true> filePath(string) => bool(true) */
 	private array $analysedFiles = [];
@@ -232,8 +167,6 @@ class NodeScopeResolver
 		private readonly Container $container,
 		private readonly ReflectionProvider $reflectionProvider,
 		private readonly InitializerExprTypeResolver $initializerExprTypeResolver,
-		private readonly Reflector $reflector,
-		private readonly ClassReflectionFactory $classReflectionFactory,
 		#[AutowiredExtensions(of: FunctionParameterOutTypeExtension::class)]
 		private readonly ExtensionsCollection $functionParameterOutTypeExtensions,
 		#[AutowiredExtensions(of: MethodParameterOutTypeExtension::class)]
@@ -245,7 +178,6 @@ class NodeScopeResolver
 		private readonly FileTypeMapper $fileTypeMapper,
 		private readonly PhpDocInheritanceResolver $phpDocInheritanceResolver,
 		private readonly FileHelper $fileHelper,
-		private readonly TypeSpecifier $typeSpecifier,
 		#[AutowiredExtensions(of: ReadWritePropertiesExtension::class)]
 		private readonly ExtensionsCollection $readWritePropertiesExtensions,
 		#[AutowiredExtensions(of: FunctionParameterClosureThisExtension::class)]
@@ -267,13 +199,10 @@ class NodeScopeResolver
 		private readonly bool $polluteScopeWithLoopInitialAssignments,
 		#[AutowiredParameter]
 		private readonly bool $polluteScopeWithAlwaysIterableForeach,
-		#[AutowiredParameter]
-		private readonly bool $polluteScopeWithBlock,
 		#[AutowiredParameter(ref: '%exceptions.implicitThrows%')]
 		private readonly bool $implicitThrows,
 		#[AutowiredParameter]
 		private readonly bool $treatPhpDocTypesAsCertain,
-		private readonly ImplicitToStringCallHelper $implicitToStringCallHelper,
 		protected readonly ExpressionResultFactory $expressionResultFactory,
 	)
 	{
@@ -395,7 +324,7 @@ class NodeScopeResolver
 	{
 	}
 
-	protected function processPendingFibers(ExpressionResultStorage $storage): void
+	public function processPendingFibers(ExpressionResultStorage $storage): void
 	{
 	}
 
@@ -552,7 +481,7 @@ class NodeScopeResolver
 	 * @param Node\Stmt[] $stmts
 	 * @param callable(Node $node, Scope $scope): void $nodeCallback
 	 */
-	private function processStmtNodesInternal(
+	public function processStmtNodesInternal(
 		Node $parentNode,
 		array $stmts,
 		MutatingScope $scope,
@@ -797,1896 +726,32 @@ class NodeScopeResolver
 
 		$this->callNodeCallback($nodeCallback, $stmt, $stmtScope, $storage);
 
-		if ($stmt instanceof Node\Stmt\Declare_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			$alwaysTerminating = false;
-			$exitPoints = [];
-			foreach ($stmt->declares as $declare) {
-				$this->callNodeCallback($nodeCallback, $declare, $scope, $storage);
-				$this->callNodeCallback($nodeCallback, $declare->value, $scope, $storage);
-				if (
-					$declare->key->name !== 'strict_types'
-					|| !($declare->value instanceof Node\Scalar\Int_)
-					|| $declare->value->value !== 1
-				) {
-					continue;
-				}
-
-				$scope = $scope->enterDeclareStrictTypes();
-			}
-
-			if ($stmt->stmts !== null) {
-				$result = $this->processStmtNodesInternal($stmt, $stmt->stmts, $scope, $storage, $nodeCallback, $context);
-				$scope = $result->getScope();
-				$hasYield = $result->hasYield();
-				$throwPoints = $result->getThrowPoints();
-				$impurePoints = $result->getImpurePoints();
-				$alwaysTerminating = $result->isAlwaysTerminating();
-				$exitPoints = $result->getExitPoints();
-			}
-
-			return new InternalStatementResult($scope, $hasYield, $alwaysTerminating, $exitPoints, $throwPoints, $impurePoints);
-		} elseif ($stmt instanceof Node\Stmt\Function_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			$this->processAttributeGroups($stmt, $stmt->attrGroups, $scope, $storage, $nodeCallback);
-			[$templateTypeMap, $phpDocParameterTypes, $phpDocImmediatelyInvokedCallableParameters, $phpDocClosureThisTypeParameters, $phpDocReturnType, $phpDocThrowType, $deprecatedDescription, $isDeprecated, $isInternal, , $isPure, $acceptsNamedArguments, , $phpDocComment, $asserts,, $phpDocParameterOutTypes, , , , $pureUnlessCallableIsImpureParameters] = $this->getPhpDocs($scope, $stmt);
-
-			foreach ($stmt->params as $param) {
-				$this->processParamNode($stmt, $param, $scope, $storage, $nodeCallback);
-			}
-
-			if ($stmt->returnType !== null) {
-				$this->callNodeCallback($nodeCallback, $stmt->returnType, $scope, $storage);
-			}
-
-			if (!$isDeprecated) {
-				[$isDeprecated, $deprecatedDescription] = $this->getDeprecatedAttribute($scope, $stmt);
-			}
-
-			$functionScope = $scope->enterFunction(
-				$stmt,
-				$templateTypeMap,
-				$phpDocParameterTypes,
-				$phpDocReturnType,
-				$phpDocThrowType,
-				$deprecatedDescription,
-				$isDeprecated,
-				$isInternal,
-				$isPure,
-				$acceptsNamedArguments,
-				$asserts,
-				$phpDocComment,
-				$phpDocParameterOutTypes,
-				$phpDocImmediatelyInvokedCallableParameters,
-				$phpDocClosureThisTypeParameters,
-				$pureUnlessCallableIsImpureParameters,
-			);
-			$functionReflection = $functionScope->getFunction();
-			if (!$functionReflection instanceof PhpFunctionFromParserNodeReflection) {
-				throw new ShouldNotHappenException();
-			}
-
-			$this->callNodeCallback($nodeCallback, new InFunctionNode($functionReflection, $stmt), $functionScope, $storage);
-
-			$gatheredReturnStatements = [];
-			$gatheredYieldStatements = [];
-			$executionEnds = [];
-			$functionImpurePoints = [];
-			$statementResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $functionScope, $storage, new GatheringNodeCallback(static function (Node $node, Scope $scope) use ($functionScope, &$gatheredReturnStatements, &$gatheredYieldStatements, &$executionEnds, &$functionImpurePoints): void {
-				if ($scope->getFunction() !== $functionScope->getFunction()) {
-					return;
-				}
-				if ($scope->isInAnonymousFunction()) {
-					return;
-				}
-				if ($node instanceof PropertyAssignNode) {
-					$functionImpurePoints[] = new ImpurePoint(
-						$scope,
-						$node,
-						'propertyAssign',
-						'property assignment',
-						true,
-					);
-					return;
-				}
-				if ($node instanceof ExecutionEndNode) {
-					$executionEnds[] = $node;
-					return;
-				}
-				if ($node instanceof Expr\Yield_ || $node instanceof Expr\YieldFrom) {
-					$gatheredYieldStatements[] = $node;
-				}
-				if (!$node instanceof Return_) {
-					return;
-				}
-
-				$gatheredReturnStatements[] = new ReturnStatement($scope, $node);
-			}, $nodeCallback), StatementContext::createTopLevel())->toPublic();
-
-			$this->callNodeCallback($nodeCallback, new FunctionReturnStatementsNode(
-				$stmt,
-				$gatheredReturnStatements,
-				$gatheredYieldStatements,
-				$statementResult,
-				$executionEnds,
-				array_merge($statementResult->getImpurePoints(), $functionImpurePoints),
-				$functionReflection,
-			), $functionScope, $storage);
-			if (!$scope->isInAnonymousFunction()) {
-				$this->processPendingFibers($storage);
-			}
-
-			// declaring the function defines it in global state, so a negative
-			// function_exists() narrowing that may refer to that function must be forgotten
-			$scope = $scope->invalidateExistenceCheckExpressions(['function_exists'], $functionReflection->getName());
-		} elseif ($stmt instanceof Node\Stmt\ClassMethod) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			$this->processAttributeGroups($stmt, $stmt->attrGroups, $scope, $storage, $nodeCallback);
-			[$templateTypeMap, $phpDocParameterTypes, $phpDocImmediatelyInvokedCallableParameters, $phpDocClosureThisTypeParameters, $phpDocReturnType, $phpDocThrowType, $deprecatedDescription, $isDeprecated, $isInternal, $isFinal, $isPure, $acceptsNamedArguments, $isReadOnly, $phpDocComment, $asserts, $selfOutType, $phpDocParameterOutTypes, , , , $pureUnlessCallableIsImpureParameters] = $this->getPhpDocs($scope, $stmt);
-
-			foreach ($stmt->params as $param) {
-				$this->processParamNode($stmt, $param, $scope, $storage, $nodeCallback);
-			}
-
-			if ($stmt->returnType !== null) {
-				$this->callNodeCallback($nodeCallback, $stmt->returnType, $scope, $storage);
-			}
-
-			if (!$isDeprecated) {
-				[$isDeprecated, $deprecatedDescription] = $this->getDeprecatedAttribute($scope, $stmt);
-			}
-
-			$isFromTrait = $stmt->getAttribute('originalTraitMethodName') === '__construct';
-			$isConstructor = $isFromTrait || $stmt->name->toLowerString() === '__construct';
-
-			$methodScope = $scope->enterClassMethod(
-				$stmt,
-				$templateTypeMap,
-				$phpDocParameterTypes,
-				$phpDocReturnType,
-				$phpDocThrowType,
-				$deprecatedDescription,
-				$isDeprecated,
-				$isInternal,
-				$isFinal,
-				$isPure,
-				$acceptsNamedArguments,
-				$asserts,
-				$selfOutType,
-				$phpDocComment,
-				$phpDocParameterOutTypes,
-				$phpDocImmediatelyInvokedCallableParameters,
-				$phpDocClosureThisTypeParameters,
-				$isConstructor,
-				null,
-				$pureUnlessCallableIsImpureParameters,
-			);
-
-			if (!$scope->isInClass()) {
-				throw new ShouldNotHappenException();
-			}
-
-			$classReflection = $scope->getClassReflection();
-
-			if ($isConstructor) {
-				foreach ($stmt->params as $param) {
-					if ($param->flags === 0 && $param->hooks === []) {
-						continue;
-					}
-
-					if (!$param->var instanceof Variable || !is_string($param->var->name) || $param->var->name === '') {
-						throw new ShouldNotHappenException();
-					}
-					$phpDoc = null;
-					if ($param->getDocComment() !== null) {
-						$phpDoc = $param->getDocComment()->getText();
-					}
-					$this->callNodeCallback($nodeCallback, new ClassPropertyNode(
-						$param->var->name,
-						$param->flags,
-						$param->type !== null ? ParserNodeTypeToPHPStanType::resolve($param->type, $classReflection) : null,
-						null,
-						$phpDoc,
-						$phpDocParameterTypes[$param->var->name] ?? null,
-						true,
-						$isFromTrait,
-						$param,
-						$isReadOnly,
-						$scope->isInTrait(),
-						$classReflection->isReadOnly(),
-						false,
-						$classReflection,
-					), $methodScope, $storage);
-					$this->processPropertyHooks(
-						$stmt,
-						$param->type,
-						$phpDocParameterTypes[$param->var->name] ?? null,
-						$param->var->name,
-						$param->hooks,
-						$scope,
-						$storage,
-						$nodeCallback,
-					);
-					$methodScope = $methodScope->assignExpression(new PropertyInitializationExpr($param->var->name), new MixedType(), new MixedType());
-				}
-			}
-
-			if ($stmt->getAttribute('virtual', false) === false) {
-				$methodReflection = $methodScope->getFunction();
-				if (!$methodReflection instanceof PhpMethodFromParserNodeReflection) {
-					throw new ShouldNotHappenException();
-				}
-				$this->callNodeCallback($nodeCallback, new InClassMethodNode($classReflection, $methodReflection, $stmt), $methodScope, $storage);
-			}
-
-			if ($stmt->stmts !== null) {
-				$gatheredReturnStatements = [];
-				$gatheredYieldStatements = [];
-				$executionEnds = [];
-				$methodImpurePoints = [];
-				$statementResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $methodScope, $storage, new GatheringNodeCallback(static function (Node $node, Scope $scope) use ($methodScope, &$gatheredReturnStatements, &$gatheredYieldStatements, &$executionEnds, &$methodImpurePoints): void {
-					if ($scope->getFunction() !== $methodScope->getFunction()) {
-						return;
-					}
-					if ($scope->isInAnonymousFunction()) {
-						return;
-					}
-					if ($node instanceof PropertyAssignNode) {
-						if (
-							$node->getPropertyFetch() instanceof Expr\PropertyFetch
-							&& $scope->getFunction() instanceof PhpMethodFromParserNodeReflection
-							&& $scope->getFunction()->getDeclaringClass()->hasConstructor()
-							&& $scope->getFunction()->getDeclaringClass()->getConstructor()->getName() === $scope->getFunction()->getName()
-							&& TypeUtils::findThisType($scope->getType($node->getPropertyFetch()->var)) !== null
-						) {
-							return;
-						}
-						$methodImpurePoints[] = new ImpurePoint(
-							$scope,
-							$node,
-							'propertyAssign',
-							'property assignment',
-							true,
-						);
-						return;
-					}
-					if ($node instanceof ExecutionEndNode) {
-						$executionEnds[] = $node;
-						return;
-					}
-					if ($node instanceof Expr\Yield_ || $node instanceof Expr\YieldFrom) {
-						$gatheredYieldStatements[] = $node;
-					}
-					if (!$node instanceof Return_) {
-						return;
-					}
-
-					$gatheredReturnStatements[] = new ReturnStatement($scope, $node);
-				}, $nodeCallback), StatementContext::createTopLevel())->toPublic();
-
-				$methodReflection = $methodScope->getFunction();
-				if (!$methodReflection instanceof PhpMethodFromParserNodeReflection) {
-					throw new ShouldNotHappenException();
-				}
-
-				$this->callNodeCallback($nodeCallback, new MethodReturnStatementsNode(
-					$stmt,
-					$gatheredReturnStatements,
-					$gatheredYieldStatements,
-					$statementResult,
-					$executionEnds,
-					array_merge($statementResult->getImpurePoints(), $methodImpurePoints),
-					$classReflection,
-					$methodReflection,
-				), $methodScope, $storage);
-
-				if ($isConstructor) {
-					$finalScope = null;
-
-					foreach ($executionEnds as $executionEnd) {
-						if ($executionEnd->getStatementResult()->isAlwaysTerminating()) {
-							continue;
-						}
-
-						$endScope = $executionEnd->getStatementResult()->getScope();
-						if ($finalScope === null) {
-							$finalScope = $endScope;
-							continue;
-						}
-
-						$finalScope = $finalScope->mergeWith($endScope);
-					}
-
-					foreach ($gatheredReturnStatements as $statement) {
-						if ($finalScope === null) {
-							$finalScope = $statement->getScope()->toMutatingScope();
-							continue;
-						}
-
-						$finalScope = $finalScope->mergeWith($statement->getScope()->toMutatingScope());
-					}
-
-					if ($finalScope !== null) {
-						$scope = $finalScope->rememberConstructorScope();
-					}
-
-				}
-			}
-			if (!$scope->getClassReflection()->isAnonymous() && !$scope->isInAnonymousFunction()) {
-				$this->processPendingFibers($storage);
-			}
-		} elseif ($stmt instanceof Echo_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			$isAlwaysTerminating = false;
-			foreach ($stmt->exprs as $echoExpr) {
-				$result = $this->processExprNode($stmt, $echoExpr, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$throwPoints = array_merge($throwPoints, $result->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $result->getImpurePoints());
-				$toStringResult = $this->implicitToStringCallHelper->processImplicitToStringCall($echoExpr, $scope);
-				$throwPoints = array_merge($throwPoints, $toStringResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $toStringResult->getImpurePoints());
-				$scope = $result->getScope();
-				$hasYield = $hasYield || $result->hasYield();
-				$isAlwaysTerminating = $isAlwaysTerminating || $result->isAlwaysTerminating();
-			}
-
-			$throwPoints = $overridingThrowPoints ?? $throwPoints;
-			$impurePoints[] = new ImpurePoint($scope, $stmt, 'echo', 'echo', true);
-			return new InternalStatementResult($scope, $hasYield, $isAlwaysTerminating, [], $throwPoints, $impurePoints);
-		} elseif ($stmt instanceof Return_) {
-			if ($stmt->expr !== null) {
-				$result = $this->processExprNode($stmt, $stmt->expr, $stmtScope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$throwPoints = $result->getThrowPoints();
-				$impurePoints = $result->getImpurePoints();
-				$scope = $result->getScope();
-				$hasYield = $result->hasYield();
-			} else {
-				$hasYield = false;
-				$throwPoints = [];
-				$impurePoints = [];
-			}
-
-			return new InternalStatementResult($scope, $hasYield, true, [
-				new InternalStatementExitPoint($stmt, $scope),
-			], $overridingThrowPoints ?? $throwPoints, $impurePoints);
-		} elseif ($stmt instanceof Continue_ || $stmt instanceof Break_) {
-			if ($stmt->num !== null) {
-				$result = $this->processExprNode($stmt, $stmt->num, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$scope = $result->getScope();
-				$hasYield = $result->hasYield();
-				$throwPoints = $result->getThrowPoints();
-				$impurePoints = $result->getImpurePoints();
-			} else {
-				$hasYield = false;
-				$throwPoints = [];
-				$impurePoints = [];
-			}
-
-			return new InternalStatementResult($scope, $hasYield, true, [
-				new InternalStatementExitPoint($stmt, $scope),
-			], $overridingThrowPoints ?? $throwPoints, $impurePoints);
-		} elseif ($stmt instanceof Goto_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-
-			return new InternalStatementResult($scope, $hasYield, true, [
-				new InternalStatementExitPoint($stmt, $scope),
-			], $overridingThrowPoints ?? $throwPoints, $impurePoints);
-		} elseif ($stmt instanceof Node\Stmt\Label) {
-			$hasYield = false;
-			$throwPoints = $overridingThrowPoints ?? [];
-			$impurePoints = [];
-		} elseif ($stmt instanceof Node\Stmt\Expression) {
-			if ($stmt->expr instanceof Expr\Throw_) {
-				$scope = $stmtScope;
-			}
-			$hasAssign = false;
-			$currentScope = $scope;
-			$result = $this->processExprNode($stmt, $stmt->expr, $scope, $storage, new GatheringNodeCallback(static function (Node $node, Scope $scope) use ($currentScope, &$hasAssign): void {
-				if (
-					!($node instanceof VariableAssignNode) && !($node instanceof PropertyAssignNode)
-					|| $scope->getAnonymousFunctionReflection() !== $currentScope->getAnonymousFunctionReflection()
-					|| $scope->getFunction() !== $currentScope->getFunction()
-				) {
-					return;
-				}
-
-				$hasAssign = true;
-			}, $nodeCallback), ExpressionContext::createTopLevel());
-			$throwPoints = array_filter($result->getThrowPoints(), static fn ($throwPoint) => $throwPoint->isExplicit());
-			if (
-				count($result->getImpurePoints()) === 0
-				&& count($throwPoints) === 0
-				&& !$stmt->expr instanceof Expr\PostInc
-				&& !$stmt->expr instanceof Expr\PreInc
-				&& !$stmt->expr instanceof Expr\PostDec
-				&& !$stmt->expr instanceof Expr\PreDec
-			) {
-				$this->callNodeCallback($nodeCallback, new NoopExpressionNode($stmt->expr, $hasAssign), $scope, $storage);
-			}
-			$scope = $result->getScope();
-			$scope = $scope->applySpecifiedTypes($this->typeSpecifier->specifyTypesInCondition(
-				$scope,
-				$stmt->expr,
-				TypeSpecifierContext::createNull(),
-			));
-			$hasYield = $result->hasYield();
-			$throwPoints = $result->getThrowPoints();
-			$impurePoints = $result->getImpurePoints();
-			$isAlwaysTerminating = $result->isAlwaysTerminating();
-
-			// The expression statement is an exit point when its value type is an
-			// explicit never: exit/die/throw, a never-returning call, or a call
-			// configured as early-terminating (the call handlers give those never).
-			// Asked on the pre-statement scope: a conditional return type must
-			// resolve against the argument types the call was made with, not
-			// against state the statement itself just changed (bug-11565).
-			$statementType = $currentScope->getType($stmt->expr);
-			if ($statementType instanceof NeverType && $statementType->isExplicit()) {
-				return new InternalStatementResult($scope, $hasYield, true, [
-					new InternalStatementExitPoint($stmt, $scope),
-				], $overridingThrowPoints ?? $throwPoints, $impurePoints);
-			}
-			return new InternalStatementResult($scope, $hasYield, $isAlwaysTerminating, [], $overridingThrowPoints ?? $throwPoints, $impurePoints);
-		} elseif ($stmt instanceof Node\Stmt\Namespace_) {
-			if ($stmt->name !== null) {
-				$scope = $scope->enterNamespace($stmt->name->toString());
-			} else {
-				$scope = $scope->enterNamespace('');
-			}
-
-			$scope = $this->processStmtNodesInternal($stmt, $stmt->stmts, $scope, $storage, $nodeCallback, $context)->getScope();
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-		} elseif ($stmt instanceof Node\Stmt\Trait_) {
-			// declaring the trait defines it in global state,
-			// so a negative trait_exists() narrowing that may refer to that trait must be forgotten
-			$name = $stmt->namespacedName ?? $stmt->name;
-			$scope = $scope->invalidateExistenceCheckExpressions(['trait_exists'], $name instanceof Name ? $name->toString() : null);
-
-			return new InternalStatementResult($scope, hasYield: false, isAlwaysTerminating: false, exitPoints: [], throwPoints: [], impurePoints: []);
-		} elseif ($stmt instanceof Node\Stmt\ClassLike) {
-			// declaring a class/interface/enum defines it in global state,
-			// so a matching negative existence-check narrowing must be forgotten
-			if ($stmt instanceof Node\Stmt\Interface_) {
-				$existenceCheckFunctionNames = ['interface_exists'];
-			} elseif ($stmt instanceof Node\Stmt\Enum_) {
-				$existenceCheckFunctionNames = ['class_exists', 'enum_exists'];
-			} else {
-				$existenceCheckFunctionNames = ['class_exists'];
-			}
-			$name = $stmt->namespacedName ?? $stmt->name;
-			$scope = $scope->invalidateExistenceCheckExpressions($existenceCheckFunctionNames, $name instanceof Name ? $name->toString() : null);
-
-			if (!$context->isTopLevel()) {
-				return new InternalStatementResult($scope, hasYield: false, isAlwaysTerminating: false, exitPoints: [], throwPoints: [], impurePoints: []);
-			}
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			if (isset($stmt->namespacedName)) {
-				$classReflection = $this->getCurrentClassReflection($stmt, $stmt->namespacedName->toString(), $scope);
-				$classScope = $scope->enterClass($classReflection);
-				$this->callNodeCallback($nodeCallback, new InClassNode($stmt, $classReflection), $classScope, $storage);
-			} elseif ($stmt instanceof Class_) {
-				if ($stmt->name === null) {
-					throw new ShouldNotHappenException();
-				}
-				if (!$stmt->isAnonymous()) {
-					$classReflection = $this->reflectionProvider->getClass($stmt->name->toString());
-				} else {
-					$classReflection = $this->reflectionProvider->getAnonymousClassReflection($stmt, $scope);
-				}
-				$classScope = $scope->enterClass($classReflection);
-				$this->callNodeCallback($nodeCallback, new InClassNode($stmt, $classReflection), $classScope, $storage);
-			} else {
-				throw new ShouldNotHappenException();
-			}
-
-			$classStatementsGatherer = new ClassStatementsGatherer($classReflection, $nodeCallback);
-			$this->processAttributeGroups($stmt, $stmt->attrGroups, $classScope, $storage, $classStatementsGatherer);
-
-			$classLikeStatements = $stmt->stmts;
-			// analyze static methods first; constructor next; instance methods and property hooks last so we can carry over the scope
-			usort($classLikeStatements, static function ($a, $b) {
-				if ($a instanceof Node\Stmt\Property) {
-					return 1;
-				}
-				if ($b instanceof Node\Stmt\Property) {
-					return -1;
-				}
-
-				if (!$a instanceof Node\Stmt\ClassMethod || !$b instanceof Node\Stmt\ClassMethod) {
-					return 0;
-				}
-
-				return [!$a->isStatic(), $a->name->toLowerString() !== '__construct'] <=> [!$b->isStatic(), $b->name->toLowerString() !== '__construct'];
-			});
-
-			$this->processStmtNodesInternal($stmt, $classLikeStatements, $classScope, $storage, $classStatementsGatherer, $context);
-			$this->callNodeCallback($nodeCallback, new ClassPropertiesNode($stmt, $this->readWritePropertiesExtensions, $classStatementsGatherer->getProperties(), $classStatementsGatherer->getPropertyUsages(), $classStatementsGatherer->getMethodCalls(), $classStatementsGatherer->getReturnStatementsNodes(), $classStatementsGatherer->getPropertyAssigns(), $classReflection), $classScope, $storage);
-			$this->callNodeCallback($nodeCallback, new ClassMethodsNode($stmt, $classStatementsGatherer->getMethods(), $classStatementsGatherer->getMethodCalls(), $classReflection), $classScope, $storage);
-			$this->callNodeCallback($nodeCallback, new ClassConstantsNode($stmt, $classStatementsGatherer->getConstants(), $classStatementsGatherer->getConstantFetches(), $classReflection), $classScope, $storage);
-			$classReflection->evictPrivateSymbols();
-			$this->calledMethodResults = [];
-		} elseif ($stmt instanceof Node\Stmt\Property) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			$this->processAttributeGroups($stmt, $stmt->attrGroups, $scope, $storage, $nodeCallback);
-
-			$nativePropertyType = $stmt->type !== null ? ParserNodeTypeToPHPStanType::resolve($stmt->type, $scope->getClassReflection()) : null;
-
-			[,,,,,,,,,,,,$isReadOnly, $docComment, ,,,$varTags, $isAllowedPrivateMutation] = $this->getPhpDocs($scope, $stmt);
-			$phpDocType = null;
-			if (isset($varTags[0]) && count($varTags) === 1) {
-				$phpDocType = $varTags[0]->getType();
-			}
-
-			foreach ($stmt->props as $prop) {
-				$this->callNodeCallback($nodeCallback, $prop, $scope, $storage);
-				if ($prop->default !== null) {
-					$this->processExprNode($stmt, $prop->default, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				}
-
-				if (!$scope->isInClass()) {
-					throw new ShouldNotHappenException();
-				}
-				$propertyName = $prop->name->toString();
-
-				if ($phpDocType === null) {
-					if (isset($varTags[$propertyName])) {
-						$phpDocType = $varTags[$propertyName]->getType();
-					}
-				}
-
-				$propStmt = clone $stmt;
-				$propStmt->setAttributes($prop->getAttributes());
-				$propStmt->setAttribute('originalPropertyStmt', $stmt);
-				$this->callNodeCallback(
-					$nodeCallback,
-					new ClassPropertyNode(
-						$propertyName,
-						$stmt->flags,
-						$nativePropertyType,
-						$prop->default,
-						$docComment,
-						$phpDocType,
-						false,
-						false,
-						$propStmt,
-						$isReadOnly,
-						$scope->isInTrait(),
-						$scope->getClassReflection()->isReadOnly(),
-						$isAllowedPrivateMutation,
-						$scope->getClassReflection(),
-					),
-					$scope,
-					$storage,
-				);
-			}
-
-			if (count($stmt->hooks) > 0) {
-				if (!isset($propertyName)) {
-					throw new ShouldNotHappenException('Property name should be known when analysing hooks.');
-				}
-				$this->processPropertyHooks(
-					$stmt,
-					$stmt->type,
-					$phpDocType,
-					$propertyName,
-					$stmt->hooks,
-					$scope,
-					$storage,
-					$nodeCallback,
-				);
-			}
-
-			if ($stmt->type !== null) {
-				$this->callNodeCallback($nodeCallback, $stmt->type, $scope, $storage);
-			}
-		} elseif ($stmt instanceof If_) {
-			$condResult = $this->processExprNode($stmt, $stmt->cond, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-			$conditionType = ($this->treatPhpDocTypesAsCertain ? $condResult->getType() : $condResult->getNativeType())->toBoolean();
-			$ifAlwaysTrue = $conditionType->isTrue()->yes();
-			$exitPoints = [];
-			$throwPoints = $overridingThrowPoints ?? $condResult->getThrowPoints();
-			$impurePoints = $condResult->getImpurePoints();
-			$endStatements = [];
-			$finalScope = null;
-			$alwaysTerminating = true;
-			$hasYield = $condResult->hasYield();
-
-			$branchScopeStatementResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $condResult->getTruthyScope(), $storage, $nodeCallback, $context);
-
-			if (!$conditionType->isTrue()->no()) {
-				$exitPoints = $branchScopeStatementResult->getExitPoints();
-				$throwPoints = array_merge($throwPoints, $branchScopeStatementResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $branchScopeStatementResult->getImpurePoints());
-				$branchScope = $branchScopeStatementResult->getScope();
-				$finalScope = $branchScopeStatementResult->isAlwaysTerminating() ? null : $branchScope;
-				$alwaysTerminating = $branchScopeStatementResult->isAlwaysTerminating();
-				if (count($branchScopeStatementResult->getEndStatements()) > 0) {
-					$endStatements = array_merge($endStatements, $branchScopeStatementResult->getEndStatements());
-				} elseif (count($stmt->stmts) > 0) {
-					$endStatements[] = new InternalEndStatementResult($stmt->stmts[count($stmt->stmts) - 1], $branchScopeStatementResult);
-				} else {
-					$endStatements[] = new InternalEndStatementResult($stmt, $branchScopeStatementResult);
-				}
-				$hasYield = $branchScopeStatementResult->hasYield() || $hasYield;
-			}
-
-			$scope = $condResult->getFalseyScope();
-			$lastElseIfConditionIsTrue = false;
-
-			$condScope = $scope;
-			foreach ($stmt->elseifs as $elseif) {
-				$this->callNodeCallback($nodeCallback, $elseif, $scope, $storage);
-				$condResult = $this->processExprNode($stmt, $elseif->cond, $condScope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$elseIfConditionType = ($this->treatPhpDocTypesAsCertain ? $condResult->getType() : $condResult->getNativeType())->toBoolean();
-				$throwPoints = array_merge($throwPoints, $condResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $condResult->getImpurePoints());
-				$branchScopeStatementResult = $this->processStmtNodesInternal($elseif, $elseif->stmts, $condResult->getTruthyScope(), $storage, $nodeCallback, $context);
-
-				if (
-					!$ifAlwaysTrue
-					&& !$lastElseIfConditionIsTrue
-					&& !$elseIfConditionType->isTrue()->no()
-				) {
-					$exitPoints = array_merge($exitPoints, $branchScopeStatementResult->getExitPoints());
-					$throwPoints = array_merge($throwPoints, $branchScopeStatementResult->getThrowPoints());
-					$impurePoints = array_merge($impurePoints, $branchScopeStatementResult->getImpurePoints());
-					$branchScope = $branchScopeStatementResult->getScope();
-					$finalScope = $branchScopeStatementResult->isAlwaysTerminating() ? $finalScope : $branchScope->mergeWith($finalScope, true);
-					$alwaysTerminating = $alwaysTerminating && $branchScopeStatementResult->isAlwaysTerminating();
-					if (count($branchScopeStatementResult->getEndStatements()) > 0) {
-						$endStatements = array_merge($endStatements, $branchScopeStatementResult->getEndStatements());
-					} elseif (count($elseif->stmts) > 0) {
-						$endStatements[] = new InternalEndStatementResult($elseif->stmts[count($elseif->stmts) - 1], $branchScopeStatementResult);
-					} else {
-						$endStatements[] = new InternalEndStatementResult($elseif, $branchScopeStatementResult);
-					}
-					$hasYield = $hasYield || $branchScopeStatementResult->hasYield();
-				}
-
-				if (
-					$elseIfConditionType->isTrue()->yes()
-				) {
-					$lastElseIfConditionIsTrue = true;
-				}
-
-				$condScope = $condResult->getFalseyScope();
-				$scope = $condScope;
-			}
-
-			if ($stmt->else === null) {
-				if (!$ifAlwaysTrue && !$lastElseIfConditionIsTrue) {
-					$finalScope = $scope->mergeWith($finalScope, true);
-					$alwaysTerminating = false;
-				}
-			} else {
-				$this->callNodeCallback($nodeCallback, $stmt->else, $scope, $storage);
-				$branchScopeStatementResult = $this->processStmtNodesInternal($stmt->else, $stmt->else->stmts, $scope, $storage, $nodeCallback, $context);
-
-				if (!$ifAlwaysTrue && !$lastElseIfConditionIsTrue) {
-					$exitPoints = array_merge($exitPoints, $branchScopeStatementResult->getExitPoints());
-					$throwPoints = array_merge($throwPoints, $branchScopeStatementResult->getThrowPoints());
-					$impurePoints = array_merge($impurePoints, $branchScopeStatementResult->getImpurePoints());
-					$branchScope = $branchScopeStatementResult->getScope();
-					$finalScope = $branchScopeStatementResult->isAlwaysTerminating() ? $finalScope : $branchScope->mergeWith($finalScope, true);
-					$alwaysTerminating = $alwaysTerminating && $branchScopeStatementResult->isAlwaysTerminating();
-					if (count($branchScopeStatementResult->getEndStatements()) > 0) {
-						$endStatements = array_merge($endStatements, $branchScopeStatementResult->getEndStatements());
-					} elseif (count($stmt->else->stmts) > 0) {
-						$endStatements[] = new InternalEndStatementResult($stmt->else->stmts[count($stmt->else->stmts) - 1], $branchScopeStatementResult);
-					} else {
-						$endStatements[] = new InternalEndStatementResult($stmt->else, $branchScopeStatementResult);
-					}
-					$hasYield = $hasYield || $branchScopeStatementResult->hasYield();
-				}
-			}
-
-			if ($finalScope === null) {
-				$finalScope = $scope;
-			}
-
-			if ($stmt->else === null && !$ifAlwaysTrue && !$lastElseIfConditionIsTrue) {
-				$endStatements[] = new InternalEndStatementResult($stmt, new InternalStatementResult($finalScope, $hasYield, $alwaysTerminating, $exitPoints, $throwPoints, $impurePoints));
-			}
-
-			return new InternalStatementResult($finalScope, $hasYield, $alwaysTerminating, $exitPoints, $throwPoints, $impurePoints, $endStatements);
-		} elseif ($stmt instanceof Node\Stmt\TraitUse) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-
-			$traitStorage = $storage->duplicate();
-			$traitStorage->pendingFibers = [];
-			$this->processTraitUse($stmt, $scope, $traitStorage, $nodeCallback);
-			$this->processPendingFibers($traitStorage);
-		} elseif ($stmt instanceof Foreach_) {
-			if ($stmt->expr instanceof Variable && is_string($stmt->expr->name)) {
-				$scope = $this->processVarAnnotation($scope, [$stmt->expr->name], $stmt);
-			}
-			$condResult = $this->processExprNode($stmt, $stmt->expr, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-			$throwPoints = $overridingThrowPoints ?? $condResult->getThrowPoints();
-			$impurePoints = $condResult->getImpurePoints();
-			$scope = $condResult->getScope();
-			$arrayComparisonExpr = new BinaryOp\NotIdentical(
-				$stmt->expr,
-				new Array_([]),
-			);
-			$this->callNodeCallback($nodeCallback, new InForeachNode($stmt), $scope, $storage);
-			$originalScope = $scope;
-			$bodyScope = $scope;
-			$foreachIterateeType = $originalScope->getType($stmt->expr);
-			$foreachNativeIterateeType = $originalScope->getNativeType($stmt->expr);
-
-			if ($stmt->keyVar instanceof Variable) {
-				$keyTypeExpr = new NativeTypeExpr(
-					$originalScope->getIterableKeyType($foreachIterateeType),
-					$originalScope->getIterableKeyType($foreachNativeIterateeType),
-				);
-				$this->callNodeCallback($nodeCallback, new VariableAssignNode($stmt->keyVar, $keyTypeExpr), $originalScope, $storage);
-			}
-
-			if ($stmt->valueVar instanceof Variable) {
-				$valueTypeExpr = new NativeTypeExpr(
-					$originalScope->getIterableValueType($foreachIterateeType),
-					$originalScope->getIterableValueType($foreachNativeIterateeType),
-				);
-				$this->callNodeCallback($nodeCallback, new VariableAssignNode($stmt->valueVar, $valueTypeExpr), $originalScope, $storage);
-			} elseif ($stmt->valueVar instanceof List_) {
-				$virtualAssign = new Assign($stmt->valueVar, new NativeTypeExpr(
-					$originalScope->getIterableValueType($foreachIterateeType),
-					$originalScope->getIterableValueType($foreachNativeIterateeType),
-				));
-				$virtualAssign->setAttributes($stmt->valueVar->getAttributes());
-				$this->callNodeCallback($nodeCallback, $virtualAssign, $scope, $storage);
-			}
-
-			$originalStorage = $storage;
-			$unrolledEndScope = null;
-			$unrolledTotalKeys = null;
-			$iterateeScope = $this->polluteScopeWithAlwaysIterableForeach ? $scope->filterByTruthyValue($arrayComparisonExpr) : $scope;
-			if ($context->isTopLevel()) {
-				$storage = $originalStorage->duplicate();
-
-				$originalScope = $iterateeScope;
-				$foreachIterateeType = $originalScope->getType($stmt->expr);
-				$foreachNativeIterateeType = $originalScope->getNativeType($stmt->expr);
-				$unrolledResult = $this->tryProcessUnrolledConstantArrayForeach($stmt, $originalScope, $originalStorage, $context, $foreachIterateeType, $foreachNativeIterateeType);
-				if ($unrolledResult !== null) {
-					$bodyScope = $unrolledResult['bodyScope'];
-					$unrolledEndScope = $unrolledResult['endScope'];
-					$unrolledTotalKeys = $unrolledResult['totalKeys'];
-				} else {
-					$bodyScope = $this->enterForeach($originalScope, $storage, $originalScope, $stmt, $foreachIterateeType, $foreachNativeIterateeType, $nodeCallback);
-					$count = 0;
-					$prevEntryScope = null;
-					do {
-						$prevScope = $bodyScope;
-						$bodyScope = $bodyScope->mergeWith($iterateeScope);
-						if ($prevEntryScope !== null && $bodyScope->equals($prevEntryScope)) {
-							// walking is deterministic in the entry scope - an unchanged entry
-							// reproduces the previous pass's exit, so the verification walk is
-							// skipped
-							$bodyScope = $prevScope;
-							break;
-						}
-						$prevEntryScope = $bodyScope;
-						$storage = $originalStorage->duplicate();
-						$bodyScope = $this->enterForeach($bodyScope, $storage, $originalScope, $stmt, $foreachIterateeType, $foreachNativeIterateeType, $nodeCallback);
-						$bodyScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, new NoopNodeCallback(), $context->enterDeep())->filterOutLoopExitPoints();
-						$bodyScope = $bodyScopeResult->getScope();
-						foreach ($bodyScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-							$bodyScope = $bodyScope->mergeWith($continueExitPoint->getScope());
-						}
-						if ($bodyScope->equals($prevScope)) {
-							break;
-						}
-
-						if ($count >= self::GENERALIZE_AFTER_ITERATION) {
-							$bodyScope = $prevScope->generalizeWith($bodyScope);
-						}
-						$count++;
-					} while ($count < self::LOOP_SCOPE_ITERATIONS);
-				}
-			}
-
-			$bodyScope = $bodyScope->mergeWith($iterateeScope);
-			$storage = $originalStorage;
-			$bodyScope = $this->enterForeach($bodyScope, $storage, $originalScope, $stmt, $foreachIterateeType, $foreachNativeIterateeType, $nodeCallback);
-			$finalPassContext = $unrolledTotalKeys !== null ? $context->enterUnrolledForeach($unrolledTotalKeys) : $context;
-			$finalScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, $nodeCallback, $finalPassContext)->filterOutLoopExitPoints();
-			$finalScope = $finalScopeResult->getScope();
-			$scopesWithIterableValueType = [];
-
-			$keyVarExpr = null;
-			$originalKeyVarExpr = null;
-			if ($stmt->keyVar instanceof Variable && is_string($stmt->keyVar->name)) {
-				$keyVarExpr = $stmt->keyVar;
-				$originalKeyVarExpr = new OriginalForeachKeyExpr($stmt->keyVar->name);
-			}
-			$originalValueExpr = null;
-			if ($stmt->valueVar instanceof Variable && is_string($stmt->valueVar->name)) {
-				$originalValueExpr = new OriginalForeachValueExpr($stmt->valueVar->name);
-			}
-
-			// With a key variable, each iteration is tracked through the original key
-			// expression and the narrowed element is projected onto the array dim fetch.
-			// Without one (`foreach ($a as $v)`) we instead track the original value
-			// expression and rewrite the array value type directly from the value var.
-			$trackingExpr = $originalKeyVarExpr ?? $originalValueExpr;
-
-			$continueExitPointHasUnoriginalKeyType = false;
-			if ($trackingExpr !== null) {
-				if ($finalScope->hasExpressionType($trackingExpr)->yes()) {
-					$scopesWithIterableValueType[] = $finalScope;
-				} else {
-					$continueExitPointHasUnoriginalKeyType = true;
-				}
-			}
-
-			foreach ($finalScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-				$continueScope = $continueExitPoint->getScope();
-				$finalScope = $continueScope->mergeWith($finalScope);
-				if ($trackingExpr === null || !$continueScope->hasExpressionType($trackingExpr)->yes()) {
-					$continueExitPointHasUnoriginalKeyType = true;
-					continue;
-				}
-				$scopesWithIterableValueType[] = $continueScope;
-			}
-			$breakExitPoints = $finalScopeResult->getExitPointsByType(Break_::class);
-			foreach ($breakExitPoints as $breakExitPoint) {
-				$finalScope = $breakExitPoint->getScope()->mergeWith($finalScope);
-			}
-
-			if ($unrolledEndScope !== null) {
-				$finalScope = $unrolledEndScope;
-			}
-
-			$exprType = $scope->getType($stmt->expr);
-			$hasExpr = $scope->hasExpressionType($stmt->expr);
-			if (
-				count($breakExitPoints) === 0
-				&& count($scopesWithIterableValueType) > 0
-				&& !$continueExitPointHasUnoriginalKeyType
-				&& ($keyVarExpr !== null || $originalValueExpr !== null)
-				&& (!$hasExpr->no() || !$stmt->expr instanceof Variable)
-				&& $exprType->isArray()->yes()
-				&& $exprType->isConstantArray()->no()
-			) {
-				$arrayDimFetchLoopTypes = [];
-				$arrayDimFetchLoopNativeTypes = [];
-				$keyLoopTypes = [];
-				$keyLoopNativeTypes = [];
-				foreach ($scopesWithIterableValueType as $scopeWithIterableValueType) {
-					if ($keyVarExpr !== null) {
-						$arrayExprDimFetch = new ArrayDimFetch($stmt->expr, $keyVarExpr);
-						$dimFetchType = $scopeWithIterableValueType->getType($arrayExprDimFetch);
-						$dimFetchNativeType = $scopeWithIterableValueType->getNativeType($arrayExprDimFetch);
-						// Condition-based narrowings like `is_string($type)` apply to the value
-						// variable but not automatically to the array dim fetch, even though the
-						// two describe the same element for a given iteration. If the value var
-						// hasn't been reassigned (OriginalForeachValueExpr still tracked) we use
-						// the narrowed value-var type in place of the broader dim fetch type so
-						// the loop's final array rewrite below picks up the sharper element type.
-						if ($originalValueExpr !== null && $scopeWithIterableValueType->hasExpressionType($originalValueExpr)->yes()) {
-							// read the loop value variable's narrowed type directly by name -
-							// it is an assigned (not processExprNode-processed) variable
-							// ($originalValueExpr !== null implies a string-named Variable)
-							$valueVarType = $scopeWithIterableValueType->getVariableType($stmt->valueVar->name);
-							if ($dimFetchType->isSuperTypeOf($valueVarType)->yes()) {
-								$dimFetchType = $valueVarType;
-							}
-							$valueVarNativeType = $scopeWithIterableValueType->getNativeType($stmt->valueVar);
-							if ($dimFetchNativeType->isSuperTypeOf($valueVarNativeType)->yes()) {
-								$dimFetchNativeType = $valueVarNativeType;
-							}
-						}
-						$keyLoopTypes[] = $scopeWithIterableValueType->getType($keyVarExpr);
-						$keyLoopNativeTypes[] = $scopeWithIterableValueType->getNativeType($keyVarExpr);
-					} else {
-						// No key variable: the narrowed value var is the array element type directly.
-						$dimFetchType = $scopeWithIterableValueType->getVariableType($stmt->valueVar->name);
-						$dimFetchNativeType = $scopeWithIterableValueType->getNativeType($stmt->valueVar);
-					}
-					$arrayDimFetchLoopTypes[] = $dimFetchType;
-					$arrayDimFetchLoopNativeTypes[] = $dimFetchNativeType;
-				}
-
-				$arrayDimFetchLoopType = TypeCombinator::union(...$arrayDimFetchLoopTypes);
-				$arrayDimFetchLoopNativeType = TypeCombinator::union(...$arrayDimFetchLoopNativeTypes);
-
-				$valueTypeChanged = !$arrayDimFetchLoopType->equals($exprType->getIterableValueType());
-				$keyTypeChanged = false;
-				$keyLoopType = $exprType->getIterableKeyType();
-				$keyLoopNativeType = $scope->getNativeType($stmt->expr)->getIterableKeyType();
-				if ($keyVarExpr !== null) {
-					$keyLoopType = TypeCombinator::union(...$keyLoopTypes);
-					$keyLoopNativeType = TypeCombinator::union(...$keyLoopNativeTypes);
-					$keyTypeChanged = !$keyLoopType->equals($exprType->getIterableKeyType());
-				}
-
-				if ($valueTypeChanged || $keyTypeChanged) {
-					$newExprType = $exprType;
-					if ($valueTypeChanged) {
-						$newExprType = $newExprType->mapValueType(static fn (Type $type): Type => $arrayDimFetchLoopType);
-					}
-					if ($keyTypeChanged) {
-						$newExprType = $newExprType->mapKeyType(static fn (Type $type): Type => $keyLoopType);
-					}
-
-					$nativeExprType = $scope->getNativeType($stmt->expr);
-					$newExprNativeType = $nativeExprType;
-					if ($valueTypeChanged) {
-						$newExprNativeType = $newExprNativeType->mapValueType(static fn (Type $type): Type => $arrayDimFetchLoopNativeType);
-					}
-					if ($keyTypeChanged) {
-						$newExprNativeType = $newExprNativeType->mapKeyType(static fn (Type $type): Type => $keyLoopNativeType);
-					}
-
-					if ($stmt->expr instanceof Variable && is_string($stmt->expr->name)) {
-						$finalScope = $finalScope->assignVariable(
-							$stmt->expr->name,
-							$newExprType,
-							$newExprNativeType,
-							$hasExpr,
-						);
-					} else {
-						$finalScope = $finalScope->assignExpression(
-							$stmt->expr,
-							$newExprType,
-							$newExprNativeType,
-						);
-					}
-				}
-			}
-
-			$isIterableAtLeastOnce = $exprType->isIterableAtLeastOnce();
-			if ($isIterableAtLeastOnce->maybe() || $exprType->isIterable()->no()) {
-				$finalScope = $finalScope->mergeWith($scope->filterByTruthyValue(new BooleanOr(
-					new BinaryOp\Identical(
-						$stmt->expr,
-						new Array_([]),
-					),
-					new FuncCall(new Name\FullyQualified('is_object'), [
-						new Arg($stmt->expr),
-					]),
-				)));
-			} elseif ($isIterableAtLeastOnce->no() || $finalScopeResult->isAlwaysTerminating()) {
-				$finalScope = $scope;
-			} elseif (!$this->polluteScopeWithAlwaysIterableForeach) {
-				$finalScope = $scope->processAlwaysIterableForeachScopeWithoutPollute($finalScope);
-				// get types from finalScope, but don't create new variables
-			}
-
-			if (!$isIterableAtLeastOnce->no()) {
-				$throwPoints = array_merge($throwPoints, $finalScopeResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $finalScopeResult->getImpurePoints());
-			}
-			$traversableThrowPoint = $this->getTraversableForeachThrowPoint($scope, $stmt->expr);
-			if ($traversableThrowPoint !== null) {
-				$throwPoints[] = $traversableThrowPoint;
-			}
-			if ($context->isTopLevel() && $stmt->byRef) {
-				$finalScope = $finalScope->assignExpression(new ForeachValueByRefExpr($stmt->valueVar), new MixedType(), new MixedType());
-			}
-
-			return new InternalStatementResult(
-				$finalScope,
-				$finalScopeResult->hasYield() || $condResult->hasYield(),
-				$isIterableAtLeastOnce->yes() && $finalScopeResult->isAlwaysTerminating(),
-				$finalScopeResult->getExitPointsForOuterLoop(),
-				$throwPoints,
-				$impurePoints,
-			);
-		} elseif ($stmt instanceof While_) {
-			$originalStorage = $storage;
-			$storage = $originalStorage->duplicate();
-			$condResult = $this->processExprNode($stmt, $stmt->cond, $scope, $storage, new NoopNodeCallback(), ExpressionContext::createDeep());
-			$beforeCondBooleanType = ($this->treatPhpDocTypesAsCertain ? $condResult->getType() : $condResult->getNativeType())->toBoolean();
-			$condScope = $condResult->getFalseyScope();
-			if (!$context->isTopLevel() && $beforeCondBooleanType->isFalse()->yes()) {
-				if (!$this->polluteScopeWithLoopInitialAssignments) {
-					$scope = $condScope->mergeWith($scope);
-				}
-
+		$stmtHandler = StmtHandlerRegistry::resolve($stmt, $this->container);
+		if ($stmtHandler !== null) {
+			$stmtResult = $stmtHandler->processStmt($this, $stmt, $stmtScope, $storage, $nodeCallback, $context);
+			if ($overridingThrowPoints !== null) {
 				return new InternalStatementResult(
-					$scope,
-					$condResult->hasYield(),
-					false,
-					[],
-					$condResult->getThrowPoints(),
-					$condResult->getImpurePoints(),
+					$stmtResult->getScope(),
+					hasYield: $stmtResult->hasYield(),
+					isAlwaysTerminating: $stmtResult->isAlwaysTerminating(),
+					exitPoints: $stmtResult->getExitPoints(),
+					throwPoints: $overridingThrowPoints,
+					impurePoints: $stmtResult->getImpurePoints(),
+					endStatements: $stmtResult->getEndStatements(),
 				);
 			}
-			$bodyScope = $condResult->getTruthyScope();
 
-			if ($context->isTopLevel()) {
-				$count = 0;
-				$prevEntryScope = null;
-				do {
-					$prevScope = $bodyScope;
-					$bodyScope = $bodyScope->mergeWith($scope);
-					if ($prevEntryScope !== null && $bodyScope->equals($prevEntryScope)) {
-						// walking is deterministic in the entry scope - an unchanged entry
-						// reproduces the previous pass's exit, so the verification walk is
-						// skipped
-						$bodyScope = $prevScope;
-						break;
-					}
-					$prevEntryScope = $bodyScope;
-					$storage = $originalStorage->duplicate();
-					$bodyScope = $this->processExprNode($stmt, $stmt->cond, $bodyScope, $storage, new NoopNodeCallback(), ExpressionContext::createDeep())->getTruthyScope();
-					$bodyScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, new NoopNodeCallback(), $context->enterDeep())->filterOutLoopExitPoints();
-					$bodyScope = $bodyScopeResult->getScope();
-					foreach ($bodyScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-						$bodyScope = $bodyScope->mergeWith($continueExitPoint->getScope());
-					}
-					if ($bodyScope->equals($prevScope)) {
-						break;
-					}
-
-					if ($count >= self::GENERALIZE_AFTER_ITERATION) {
-						$bodyScope = $prevScope->generalizeWith($bodyScope);
-					}
-					$count++;
-				} while ($count < self::LOOP_SCOPE_ITERATIONS);
-			}
-
-			$bodyScope = $bodyScope->mergeWith($scope);
-			$bodyScopeMaybeRan = $bodyScope;
-			$storage = $originalStorage;
-			$bodyScope = $this->processExprNode($stmt, $stmt->cond, $bodyScope, $storage, $nodeCallback, ExpressionContext::createDeep())->getTruthyScope();
-			$finalScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, $nodeCallback, $context)->filterOutLoopExitPoints();
-			$finalScope = $finalScopeResult->getScope()->filterByFalseyValue($stmt->cond);
-
-			$alwaysIterates = false;
-			$neverIterates = false;
-			if ($context->isTopLevel()) {
-				$condBooleanType = ($this->treatPhpDocTypesAsCertain ? $bodyScopeMaybeRan->getType($stmt->cond) : $bodyScopeMaybeRan->getNativeType($stmt->cond))->toBoolean();
-				$alwaysIterates = $condBooleanType->isTrue()->yes();
-				$neverIterates = $condBooleanType->isFalse()->yes();
-			}
-			if (!$alwaysIterates) {
-				foreach ($finalScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-					$finalScope = $finalScope->mergeWith($continueExitPoint->getScope());
-				}
-			}
-
-			$breakExitPoints = $finalScopeResult->getExitPointsByType(Break_::class);
-			if (count($breakExitPoints) > 0) {
-				$breakScope = $alwaysIterates ? null : $finalScope;
-				foreach ($breakExitPoints as $breakExitPoint) {
-					$breakScope = $breakScope === null ? $breakExitPoint->getScope() : $breakScope->mergeWith($breakExitPoint->getScope());
-				}
-				$finalScope = $breakScope;
-			}
-
-			$isIterableAtLeastOnce = $beforeCondBooleanType->isTrue()->yes();
-			$this->callNodeCallback($nodeCallback, new BreaklessWhileLoopNode($stmt, $finalScopeResult->toPublic()->getExitPoints(), $finalScopeResult->hasYield()), $bodyScopeMaybeRan, $storage);
-
-			if ($alwaysIterates) {
-				$isAlwaysTerminating = count($finalScopeResult->getExitPointsByType(Break_::class)) === 0;
-			} elseif ($isIterableAtLeastOnce) {
-				$isAlwaysTerminating = $finalScopeResult->isAlwaysTerminating();
-			} else {
-				$isAlwaysTerminating = false;
-			}
-			if (!$isIterableAtLeastOnce) {
-				if (!$this->polluteScopeWithLoopInitialAssignments) {
-					$condScope = $condScope->mergeWith($scope);
-				}
-				$finalScope = $finalScope->mergeWith($condScope);
-			}
-
-			$throwPoints = $overridingThrowPoints ?? $condResult->getThrowPoints();
-			$impurePoints = $condResult->getImpurePoints();
-			if (!$neverIterates) {
-				$throwPoints = array_merge($throwPoints, $finalScopeResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $finalScopeResult->getImpurePoints());
-			}
-
-			return new InternalStatementResult(
-				$finalScope,
-				$finalScopeResult->hasYield() || $condResult->hasYield(),
-				$isAlwaysTerminating,
-				$finalScopeResult->getExitPointsForOuterLoop(),
-				$throwPoints,
-				$impurePoints,
-			);
-		} elseif ($stmt instanceof Do_) {
-			$finalScope = null;
-			$bodyScope = $scope;
-			$count = 0;
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			$originalStorage = $storage;
-
-			if ($context->isTopLevel()) {
-				$prevEntryScope = null;
-				do {
-					$prevScope = $bodyScope;
-					$bodyScope = $bodyScope->mergeWith($scope);
-					if ($prevEntryScope !== null && $bodyScope->equals($prevEntryScope)) {
-						// walking is deterministic in the entry scope - an unchanged entry
-						// reproduces the previous pass's exit, so the verification walk is
-						// skipped (and repeats only idempotent merges into the final scope)
-						$bodyScope = $prevScope;
-						break;
-					}
-					$prevEntryScope = $bodyScope;
-					$storage = $originalStorage->duplicate();
-					$bodyScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, new NoopNodeCallback(), $context->enterDeep())->filterOutLoopExitPoints();
-					$alwaysTerminating = $bodyScopeResult->isAlwaysTerminating();
-					$bodyScope = $bodyScopeResult->getScope();
-					foreach ($bodyScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-						$bodyScope = $bodyScope->mergeWith($continueExitPoint->getScope());
-					}
-					$finalScope = $alwaysTerminating ? $finalScope : $bodyScope->mergeWith($finalScope);
-					foreach ($bodyScopeResult->getExitPointsByType(Break_::class) as $breakExitPoint) {
-						$finalScope = $breakExitPoint->getScope()->mergeWith($finalScope);
-					}
-					$bodyScope = $this->processExprNode($stmt, $stmt->cond, $bodyScope, $storage, new NoopNodeCallback(), ExpressionContext::createDeep())->getTruthyScope();
-					if ($bodyScope->equals($prevScope)) {
-						break;
-					}
-
-					if ($count >= self::GENERALIZE_AFTER_ITERATION) {
-						$bodyScope = $prevScope->generalizeWith($bodyScope);
-					}
-					$count++;
-				} while ($count < self::LOOP_SCOPE_ITERATIONS);
-
-				$bodyScope = $bodyScope->mergeWith($scope);
-			}
-
-			$storage = $originalStorage;
-			$bodyScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, $nodeCallback, $context)->filterOutLoopExitPoints();
-			$bodyScope = $bodyScopeResult->getScope();
-			foreach ($bodyScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-				$bodyScope = $bodyScope->mergeWith($continueExitPoint->getScope());
-			}
-
-			$alwaysIterates = false;
-			if ($context->isTopLevel()) {
-				$condBooleanType = ($this->treatPhpDocTypesAsCertain ? $bodyScope->getType($stmt->cond) : $bodyScope->getNativeType($stmt->cond))->toBoolean();
-				$alwaysIterates = $condBooleanType->isTrue()->yes();
-			}
-
-			$this->callNodeCallback($nodeCallback, new DoWhileLoopConditionNode($stmt->cond, $bodyScopeResult->toPublic()->getExitPoints(), $bodyScopeResult->hasYield()), $bodyScope, $storage);
-
-			if ($alwaysIterates) {
-				$alwaysTerminating = count($bodyScopeResult->getExitPointsByType(Break_::class)) === 0;
-			} else {
-				$alwaysTerminating = $bodyScopeResult->isAlwaysTerminating();
-			}
-			$finalScope = $alwaysTerminating ? $finalScope : $bodyScope->mergeWith($finalScope);
-			if ($finalScope === null) {
-				$finalScope = $scope;
-			}
-			if (!$alwaysTerminating) {
-				$condResult = $this->processExprNode($stmt, $stmt->cond, $bodyScope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$hasYield = $condResult->hasYield();
-				$throwPoints = $condResult->getThrowPoints();
-				$impurePoints = $condResult->getImpurePoints();
-				$finalScope = $condResult->getFalseyScope();
-			} else {
-				$this->processExprNode($stmt, $stmt->cond, $bodyScope, $storage, $nodeCallback, ExpressionContext::createDeep());
-			}
-
-			$breakExitPoints = $bodyScopeResult->getExitPointsByType(Break_::class);
-			if (count($breakExitPoints) > 0) {
-				$breakScope = $alwaysIterates ? null : $finalScope;
-				foreach ($breakExitPoints as $breakExitPoint) {
-					$breakScope = $breakScope === null ? $breakExitPoint->getScope() : $breakScope->mergeWith($breakExitPoint->getScope());
-				}
-				$finalScope = $breakScope;
-			}
-
-			return new InternalStatementResult(
-				$finalScope,
-				$bodyScopeResult->hasYield() || $hasYield,
-				$alwaysTerminating,
-				$bodyScopeResult->getExitPointsForOuterLoop(),
-				array_merge($throwPoints, $bodyScopeResult->getThrowPoints()),
-				array_merge($impurePoints, $bodyScopeResult->getImpurePoints()),
-			);
-		} elseif ($stmt instanceof For_) {
-			$initScope = $scope;
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			foreach ($stmt->init as $initExpr) {
-				$initResult = $this->processExprNode($stmt, $initExpr, $initScope, $storage, $nodeCallback, ExpressionContext::createTopLevel());
-				$initScope = $initResult->getScope();
-				$hasYield = $hasYield || $initResult->hasYield();
-				$throwPoints = array_merge($throwPoints, $initResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $initResult->getImpurePoints());
-			}
-
-			$originalStorage = $storage;
-
-			$bodyScope = $initScope;
-			$isIterableAtLeastOnce = TrinaryLogic::createYes();
-			$lastCondExpr = array_last($stmt->cond);
-			if (count($stmt->cond) > 0) {
-				$storage = $originalStorage->duplicate();
-
-				foreach ($stmt->cond as $condExpr) {
-					$condResult = $this->processExprNode($stmt, $condExpr, $bodyScope, $storage, new NoopNodeCallback(), ExpressionContext::createDeep());
-					$initScope = $condResult->getScope();
-
-					// only the last condition expression is relevant whether the loop continues
-					// see https://www.php.net/manual/en/control-structures.for.php
-					if ($condExpr === $lastCondExpr) {
-						$condTruthiness = ($this->treatPhpDocTypesAsCertain ? $condResult->getType() : $condResult->getNativeType())->toBoolean();
-						$isIterableAtLeastOnce = $isIterableAtLeastOnce->and($condTruthiness->isTrue());
-					}
-
-					$hasYield = $hasYield || $condResult->hasYield();
-					$throwPoints = array_merge($throwPoints, $condResult->getThrowPoints());
-					$impurePoints = array_merge($impurePoints, $condResult->getImpurePoints());
-					$bodyScope = $condResult->getTruthyScope();
-				}
-			}
-
-			if ($context->isTopLevel()) {
-				$count = 0;
-				$prevEntryScope = null;
-				do {
-					$prevScope = $bodyScope;
-					$storage = $originalStorage->duplicate();
-					$bodyScope = $bodyScope->mergeWith($initScope);
-					if ($prevEntryScope !== null && $bodyScope->equals($prevEntryScope)) {
-						// walking is deterministic in the entry scope - an unchanged entry
-						// reproduces the previous pass's exit, so the verification walk is
-						// skipped
-						$bodyScope = $prevScope;
-						break;
-					}
-					$prevEntryScope = $bodyScope;
-					if ($lastCondExpr !== null) {
-						$bodyScope = $this->processExprNode($stmt, $lastCondExpr, $bodyScope, $storage, new NoopNodeCallback(), ExpressionContext::createDeep())->getTruthyScope();
-					}
-					$bodyScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, new NoopNodeCallback(), $context->enterDeep())->filterOutLoopExitPoints();
-					$bodyScope = $bodyScopeResult->getScope();
-					foreach ($bodyScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-						$bodyScope = $bodyScope->mergeWith($continueExitPoint->getScope());
-					}
-
-					foreach ($stmt->loop as $loopExpr) {
-						$exprResult = $this->processExprNode($stmt, $loopExpr, $bodyScope, $storage, new NoopNodeCallback(), ExpressionContext::createTopLevel());
-						$bodyScope = $exprResult->getScope();
-						$hasYield = $hasYield || $exprResult->hasYield();
-						$throwPoints = array_merge($throwPoints, $exprResult->getThrowPoints());
-						$impurePoints = array_merge($impurePoints, $exprResult->getImpurePoints());
-					}
-
-					if ($bodyScope->equals($prevScope)) {
-						break;
-					}
-
-					if ($count >= self::GENERALIZE_AFTER_ITERATION) {
-						$bodyScope = $prevScope->generalizeWith($bodyScope);
-					}
-					$count++;
-				} while ($count < self::LOOP_SCOPE_ITERATIONS);
-			}
-
-			$storage = $originalStorage;
-			$bodyScope = $bodyScope->mergeWith($initScope);
-
-			$alwaysIterates = TrinaryLogic::createFromBoolean($context->isTopLevel());
-			if ($lastCondExpr !== null) {
-				$alwaysIterates = $alwaysIterates->and($bodyScope->getType($lastCondExpr)->toBoolean()->isTrue());
-				$bodyScope = $this->processExprNode($stmt, $lastCondExpr, $bodyScope, $storage, $nodeCallback, ExpressionContext::createDeep())->getTruthyScope();
-				$bodyScope = $this->inferForLoopExpressions($stmt, $lastCondExpr, $bodyScope);
-			}
-
-			$finalScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $bodyScope, $storage, $nodeCallback, $context)->filterOutLoopExitPoints();
-			$finalScope = $finalScopeResult->getScope();
-			foreach ($finalScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-				$finalScope = $continueExitPoint->getScope()->mergeWith($finalScope);
-			}
-
-			$loopScope = $finalScope;
-			foreach ($stmt->loop as $loopExpr) {
-				$loopScope = $this->processExprNode($stmt, $loopExpr, $loopScope, $storage, $nodeCallback, ExpressionContext::createTopLevel())->getScope();
-			}
-			$finalScope = $finalScope->generalizeWith($loopScope);
-
-			if ($lastCondExpr !== null) {
-				$finalScope = $finalScope->filterByFalseyValue($lastCondExpr);
-			}
-
-			$breakExitPoints = $finalScopeResult->getExitPointsByType(Break_::class);
-			if (count($breakExitPoints) > 0) {
-				$breakScope = $alwaysIterates->yes() ? null : $finalScope;
-				foreach ($breakExitPoints as $breakExitPoint) {
-					$breakScope = $breakScope === null ? $breakExitPoint->getScope() : $breakScope->mergeWith($breakExitPoint->getScope());
-				}
-				$finalScope = $breakScope;
-			}
-
-			if ($isIterableAtLeastOnce->no() || $finalScopeResult->isAlwaysTerminating()) {
-				if ($this->polluteScopeWithLoopInitialAssignments) {
-					$finalScope = $initScope;
-				} else {
-					$finalScope = $scope;
-				}
-
-			} elseif ($isIterableAtLeastOnce->maybe()) {
-				if ($this->polluteScopeWithLoopInitialAssignments) {
-					$finalScope = $finalScope->mergeWith($initScope);
-				} else {
-					$finalScope = $finalScope->mergeWith($scope);
-				}
-			} else {
-				if (!$this->polluteScopeWithLoopInitialAssignments) {
-					$finalScope = $finalScope->mergeWith($scope);
-				}
-			}
-
-			if ($alwaysIterates->yes()) {
-				$isAlwaysTerminating = count($finalScopeResult->getExitPointsByType(Break_::class)) === 0;
-			} elseif ($isIterableAtLeastOnce->yes()) {
-				$isAlwaysTerminating = $finalScopeResult->isAlwaysTerminating();
-			} else {
-				$isAlwaysTerminating = false;
-			}
-
-			return new InternalStatementResult(
-				$finalScope,
-				$finalScopeResult->hasYield() || $hasYield,
-				$isAlwaysTerminating,
-				$finalScopeResult->getExitPointsForOuterLoop(),
-				array_merge($throwPoints, $finalScopeResult->getThrowPoints()),
-				array_merge($impurePoints, $finalScopeResult->getImpurePoints()),
-			);
-		} elseif ($stmt instanceof Switch_) {
-			$condResult = $this->processExprNode($stmt, $stmt->cond, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-			$scope = $condResult->getScope();
-			$scopeForBranches = $scope;
-			$finalScope = null;
-			$prevScope = null;
-			$hasDefaultCase = false;
-			$alwaysTerminating = true;
-			$hasYield = $condResult->hasYield();
-			$exitPointsForOuterLoop = [];
-			$throwPoints = $condResult->getThrowPoints();
-			$impurePoints = $condResult->getImpurePoints();
-			$fullCondExpr = null;
-			$switchConditionArms = [];
-			$lastNonDefaultCaseKey = null;
-			foreach ($stmt->cases as $caseKey => $caseNode) {
-				if ($caseNode->cond === null) {
-					continue;
-				}
-
-				$lastNonDefaultCaseKey = $caseKey;
-			}
-			foreach ($stmt->cases as $caseKey => $caseNode) {
-				if ($caseNode->cond !== null) {
-					$condExpr = new BinaryOp\Equal($stmt->cond, $caseNode->cond);
-					$fullCondExpr = $fullCondExpr === null ? $condExpr : new BooleanOr($fullCondExpr, $condExpr);
-					$caseResult = $this->processExprNode($stmt, $caseNode->cond, $scopeForBranches, $storage, $nodeCallback, ExpressionContext::createDeep());
-					$scopeForBranches = $caseResult->getScope();
-					$hasYield = $hasYield || $caseResult->hasYield();
-					$throwPoints = array_merge($throwPoints, $caseResult->getThrowPoints());
-					$impurePoints = array_merge($impurePoints, $caseResult->getImpurePoints());
-					$switchConditionArms[] = new SwitchConditionArm(
-						$caseNode->cond,
-						$scopeForBranches,
-						$caseNode->cond->getStartLine(),
-						$caseKey === $lastNonDefaultCaseKey,
-					);
-					$branchScope = $caseResult->getScope()->filterByTruthyValue($condExpr);
-				} else {
-					$hasDefaultCase = true;
-					$fullCondExpr = null;
-					$branchScope = $scopeForBranches;
-				}
-
-				$branchScope = $branchScope->mergeWith($prevScope);
-				$branchScopeResult = $this->processStmtNodesInternal($caseNode, $caseNode->stmts, $branchScope, $storage, $nodeCallback, $context);
-				$branchScope = $branchScopeResult->getScope();
-				$branchFinalScopeResult = $branchScopeResult->filterOutLoopExitPoints();
-				$hasYield = $hasYield || $branchFinalScopeResult->hasYield();
-				foreach ($branchScopeResult->getExitPointsByType(Break_::class) as $breakExitPoint) {
-					$alwaysTerminating = false;
-					$finalScope = $breakExitPoint->getScope()->mergeWith($finalScope);
-				}
-				foreach ($branchScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-					$finalScope = $continueExitPoint->getScope()->mergeWith($finalScope);
-				}
-				$exitPointsForOuterLoop = array_merge($exitPointsForOuterLoop, $branchFinalScopeResult->getExitPointsForOuterLoop());
-				$throwPoints = array_merge($throwPoints, $branchFinalScopeResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $branchFinalScopeResult->getImpurePoints());
-				if ($branchScopeResult->isAlwaysTerminating()) {
-					$alwaysTerminating = $alwaysTerminating && $branchFinalScopeResult->isAlwaysTerminating();
-					$prevScope = null;
-					if (isset($fullCondExpr)) {
-						$scopeForBranches = $scopeForBranches->filterByFalseyValue($fullCondExpr);
-						$fullCondExpr = null;
-					}
-					if (!$branchFinalScopeResult->isAlwaysTerminating()) {
-						$finalScope = $branchScope->mergeWith($finalScope);
-					}
-				} else {
-					$prevScope = $branchScope;
-				}
-			}
-
-			if ($switchConditionArms !== []) {
-				$this->callNodeCallback($nodeCallback, new SwitchConditionNode($stmt->cond, $switchConditionArms, $stmt), $scope, $storage);
-			}
-
-			$exhaustive = $scopeForBranches->getType($stmt->cond) instanceof NeverType;
-
-			if (!$hasDefaultCase && !$exhaustive) {
-				$alwaysTerminating = false;
-			}
-
-			if ($prevScope !== null && isset($branchFinalScopeResult)) {
-				$finalScope = $prevScope->mergeWith($finalScope);
-				$alwaysTerminating = $alwaysTerminating && $branchFinalScopeResult->isAlwaysTerminating();
-			}
-
-			if ((!$hasDefaultCase && !$exhaustive) || $finalScope === null) {
-				$finalScope = $scopeForBranches->mergeWith($finalScope);
-			}
-
-			return new InternalStatementResult($finalScope, $hasYield, $alwaysTerminating, $exitPointsForOuterLoop, $throwPoints, $impurePoints);
-		} elseif ($stmt instanceof TryCatch) {
-			$branchScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $scope, $storage, $nodeCallback, $context);
-			$branchScope = $branchScopeResult->getScope();
-			$finalScope = $branchScopeResult->isAlwaysTerminating() ? null : $branchScope;
-
-			$exitPoints = [];
-			$finallyExitPoints = [];
-			$alwaysTerminating = $branchScopeResult->isAlwaysTerminating();
-			$hasYield = $branchScopeResult->hasYield();
-
-			if ($stmt->finally !== null) {
-				$finallyScope = $branchScope;
-			} else {
-				$finallyScope = null;
-			}
-			foreach ($branchScopeResult->getExitPoints() as $exitPoint) {
-				$finallyExitPoints[] = $exitPoint->toPublic();
-				if ($exitPoint->getStatement() instanceof Node\Stmt\Expression && $exitPoint->getStatement()->expr instanceof Expr\Throw_) {
-					continue;
-				}
-				if ($finallyScope !== null) {
-					$finallyScope = $finallyScope->mergeWith($exitPoint->getScope());
-				}
-				$exitPoints[] = $exitPoint;
-			}
-
-			$throwPoints = $branchScopeResult->getThrowPoints();
-			$impurePoints = $branchScopeResult->getImpurePoints();
-			$throwPointsForLater = [];
-			$pastCatchTypes = new NeverType();
-
-			foreach ($stmt->catches as $catchNode) {
-				$this->callNodeCallback($nodeCallback, $catchNode, $scope, $storage);
-
-				$originalCatchTypes = [];
-				$catchTypes = [];
-				foreach ($catchNode->types as $catchNodeType) {
-					$catchType = new ObjectType($catchNodeType->toString());
-					$originalCatchTypes[] = $catchType;
-					$catchTypes[] = TypeCombinator::remove($catchType, $pastCatchTypes);
-				}
-
-				$originalCatchType = TypeCombinator::union(...$originalCatchTypes);
-				$catchType = TypeCombinator::union(...$catchTypes);
-				$pastCatchTypes = TypeCombinator::union($pastCatchTypes, $originalCatchType);
-
-				$matchingThrowPoints = [];
-				$matchingCatchTypes = array_fill_keys(array_keys($originalCatchTypes), false);
-
-				// throwable matches all
-				foreach ($originalCatchTypes as $catchTypeIndex => $catchTypeItem) {
-					if (!$catchTypeItem->isSuperTypeOf(new ObjectType(Throwable::class))->yes()) {
-						continue;
-					}
-
-					foreach ($throwPoints as $throwPointIndex => $throwPoint) {
-						$matchingThrowPoints[$throwPointIndex] = $throwPoint;
-						$matchingCatchTypes[$catchTypeIndex] = true;
-					}
-				}
-
-				// explicit only
-				$onlyExplicitIsThrow = true;
-				if (count($matchingThrowPoints) === 0) {
-					foreach ($throwPoints as $throwPointIndex => $throwPoint) {
-						foreach ($catchTypes as $catchTypeIndex => $catchTypeItem) {
-							if ($catchTypeItem->isSuperTypeOf($throwPoint->getType())->no()) {
-								continue;
-							}
-
-							$matchingCatchTypes[$catchTypeIndex] = true;
-							if (!$throwPoint->isExplicit()) {
-								continue;
-							}
-							$throwNode = $throwPoint->getNode();
-							if (
-								!$throwNode instanceof Expr\Throw_
-								&& !($throwNode instanceof Node\Stmt\Expression && $throwNode->expr instanceof Expr\Throw_)
-							) {
-								$onlyExplicitIsThrow = false;
-							}
-
-							$matchingThrowPoints[$throwPointIndex] = $throwPoint;
-						}
-					}
-				}
-
-				// implicit only
-				if (count($matchingThrowPoints) === 0 || $onlyExplicitIsThrow) {
-					foreach ($throwPoints as $throwPointIndex => $throwPoint) {
-						if ($throwPoint->isExplicit()) {
-							continue;
-						}
-
-						foreach ($catchTypes as $catchTypeItem) {
-							if ($catchTypeItem->isSuperTypeOf($throwPoint->getType())->no()) {
-								continue;
-							}
-
-							$matchingThrowPoints[$throwPointIndex] = $throwPoint;
-						}
-					}
-				}
-
-				// include previously removed throw points
-				if (count($matchingThrowPoints) === 0) {
-					if ($originalCatchType->isSuperTypeOf(new ObjectType(Throwable::class))->yes()) {
-						foreach ($branchScopeResult->getThrowPoints() as $originalThrowPoint) {
-							if (!$originalThrowPoint->canContainAnyThrowable()) {
-								continue;
-							}
-
-							$matchingThrowPoints[] = $originalThrowPoint;
-							$matchingCatchTypes = array_fill_keys(array_keys($originalCatchTypes), true);
-						}
-					}
-				}
-
-				// emit error
-				foreach ($matchingCatchTypes as $catchTypeIndex => $matched) {
-					if ($matched) {
-						continue;
-					}
-					$this->callNodeCallback($nodeCallback, new CatchWithUnthrownExceptionNode($catchNode, $catchTypes[$catchTypeIndex], $originalCatchTypes[$catchTypeIndex]), $scope, $storage);
-				}
-
-				if (count($matchingThrowPoints) === 0) {
-					continue;
-				}
-
-				// recompute throw points
-				$newThrowPoints = [];
-				foreach ($throwPoints as $throwPoint) {
-					$newThrowPoint = $throwPoint->subtractCatchType($originalCatchType);
-
-					if ($newThrowPoint->getType() instanceof NeverType) {
-						continue;
-					}
-
-					$newThrowPoints[] = $newThrowPoint;
-				}
-				$throwPoints = $newThrowPoints;
-
-				$catchScope = null;
-				foreach ($matchingThrowPoints as $matchingThrowPoint) {
-					if ($catchScope === null) {
-						$catchScope = $matchingThrowPoint->getScope();
-					} else {
-						$catchScope = $catchScope->mergeWith($matchingThrowPoint->getScope());
-					}
-				}
-
-				$variableName = null;
-				if ($catchNode->var !== null) {
-					if (!is_string($catchNode->var->name)) {
-						throw new ShouldNotHappenException();
-					}
-
-					$variableName = $catchNode->var->name;
-					$this->callNodeCallback($nodeCallback, new VariableAssignNode($catchNode->var, new TypeExpr($catchType)), $scope, $storage);
-				}
-
-				$catchScopeResult = $this->processStmtNodesInternal($catchNode, $catchNode->stmts, $catchScope->enterCatchType($catchType, $variableName), $storage, $nodeCallback, $context);
-				$catchScopeForFinally = $catchScopeResult->getScope();
-
-				$finalScope = $catchScopeResult->isAlwaysTerminating() ? $finalScope : $catchScopeResult->getScope()->mergeWith($finalScope);
-				$alwaysTerminating = $alwaysTerminating && $catchScopeResult->isAlwaysTerminating();
-				$hasYield = $hasYield || $catchScopeResult->hasYield();
-				$catchThrowPoints = $catchScopeResult->getThrowPoints();
-				$impurePoints = array_merge($impurePoints, $catchScopeResult->getImpurePoints());
-				$throwPointsForLater = array_merge($throwPointsForLater, $catchThrowPoints);
-
-				if ($finallyScope !== null) {
-					$finallyScope = $finallyScope->mergeWith($catchScopeForFinally);
-				}
-				foreach ($catchScopeResult->getExitPoints() as $exitPoint) {
-					$finallyExitPoints[] = $exitPoint->toPublic();
-					if ($exitPoint->getStatement() instanceof Node\Stmt\Expression && $exitPoint->getStatement()->expr instanceof Expr\Throw_) {
-						continue;
-					}
-					if ($finallyScope !== null) {
-						$finallyScope = $finallyScope->mergeWith($exitPoint->getScope());
-					}
-					$exitPoints[] = $exitPoint;
-				}
-
-				foreach ($catchThrowPoints as $catchThrowPoint) {
-					if ($finallyScope === null) {
-						continue;
-					}
-					$finallyScope = $finallyScope->mergeWith($catchThrowPoint->getScope());
-				}
-			}
-
-			if ($finalScope === null) {
-				$finalScope = $scope;
-			}
-
-			foreach ($throwPoints as $throwPoint) {
-				if ($finallyScope === null) {
-					continue;
-				}
-				$finallyScope = $finallyScope->mergeWith($throwPoint->getScope());
-			}
-
-			if ($finallyScope !== null) {
-				$originalFinallyScope = $finallyScope;
-				$finallyResult = $this->processStmtNodesInternal($stmt->finally, $stmt->finally->stmts, $finallyScope, $storage, $nodeCallback, $context);
-				$alwaysTerminating = $alwaysTerminating || $finallyResult->isAlwaysTerminating();
-				$hasYield = $hasYield || $finallyResult->hasYield();
-				$throwPointsForLater = array_merge($throwPointsForLater, $finallyResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $finallyResult->getImpurePoints());
-				$finallyScope = $finallyResult->getScope();
-				$finalScope = $finallyResult->isAlwaysTerminating() ? $finalScope : $finalScope->processFinallyScope($finallyScope, $originalFinallyScope);
-				if (count($finallyResult->getExitPoints()) > 0 && $finallyResult->isAlwaysTerminating()) {
-					$this->callNodeCallback($nodeCallback, new FinallyExitPointsNode(
-						$finallyResult->toPublic()->getExitPoints(),
-						$finallyExitPoints,
-					), $scope, $storage);
-				}
-				$exitPoints = array_merge($exitPoints, $finallyResult->getExitPoints());
-			}
-
-			return new InternalStatementResult($finalScope, $hasYield, $alwaysTerminating, $exitPoints, array_merge($throwPoints, $throwPointsForLater), $impurePoints);
-		} elseif ($stmt instanceof Unset_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			foreach ($stmt->vars as $var) {
-				$scope = $this->lookForSetAllowedUndefinedExpressions($scope, $var);
-				$exprResult = $this->processExprNode($stmt, $var, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$scope = $exprResult->getScope();
-				$scope = $this->lookForUnsetAllowedUndefinedExpressions($scope, $var);
-				$hasYield = $hasYield || $exprResult->hasYield();
-				$throwPoints = array_merge($throwPoints, $exprResult->getThrowPoints());
-				$impurePoints = array_merge($impurePoints, $exprResult->getImpurePoints());
-				if ($var instanceof ArrayDimFetch && $var->dim !== null) {
-					$varType = $scope->getType($var->var);
-					if (!$varType->isArray()->yes() && !(new ObjectType(ArrayAccess::class))->isSuperTypeOf($varType)->no()) {
-						$throwPoints = array_merge($throwPoints, $this->container->getByType(MethodThrowPointHelper::class)->getThrowPointsForCallOnType(
-							$scope,
-							ExpressionContext::createDeep(),
-							$varType,
-							new MethodCall(new TypeExpr($varType), 'offsetUnset'),
-						));
-					}
-
-					// wrap the already-processed chain in ExistingArrayDimFetch nodes
-					// referencing the original sub-expressions - the unset statement's
-					// own walk already processed them, and the assign target
-					// preparation prices the chain without re-walking it
-					$buildExistingChain = static function (Expr $node) use (&$buildExistingChain): Expr {
-						if (!$node instanceof ArrayDimFetch || $node->dim === null) {
-							return $node;
-						}
-
-						return new ExistingArrayDimFetch(
-							$buildExistingChain($node->var),
-							$node->dim,
-						);
-					};
-					$scope = $this->processVirtualAssign($scope, $storage, $stmt, $buildExistingChain($var->var), new UnsetOffsetExpr($var->var, $var->dim), $nodeCallback)->getScope();
-				} elseif ($var instanceof PropertyFetch) {
-					$scope = $scope->invalidateExpression($var);
-					$impurePoints[] = new ImpurePoint(
-						$scope,
-						$var,
-						'propertyUnset',
-						'property unset',
-						true,
-					);
-				} else {
-					$scope = $scope->invalidateExpression($var);
-				}
-
-				$scope = $scope->invalidateExpression(new ForeachValueByRefExpr($var));
-			}
-		} elseif ($stmt instanceof Node\Stmt\Use_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			foreach ($stmt->uses as $use) {
-				$this->callNodeCallback($nodeCallback, $use, $scope, $storage);
-			}
-		} elseif ($stmt instanceof Node\Stmt\Global_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [
-				new ImpurePoint(
-					$scope,
-					$stmt,
-					'global',
-					'global variable',
-					true,
-				),
-			];
-			$vars = [];
-			foreach ($stmt->vars as $var) {
-				if (!$var instanceof Variable) {
-					throw new ShouldNotHappenException();
-				}
-				$scope = $this->lookForSetAllowedUndefinedExpressions($scope, $var);
-				$varResult = $this->processExprNode($stmt, $var, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$impurePoints = array_merge($impurePoints, $varResult->getImpurePoints());
-				$scope = $this->lookForUnsetAllowedUndefinedExpressions($scope, $var);
-
-				if (!is_string($var->name)) {
-					continue;
-				}
-
-				$varType = $this->getGlobalVariableType($var->name);
-				$scope = $scope->assignVariable($var->name, $varType, $varType, TrinaryLogic::createYes());
-				$vars[] = $var->name;
-			}
-			$scope = $this->processVarAnnotation($scope, $vars, $stmt);
-		} elseif ($stmt instanceof Static_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [
-				new ImpurePoint(
-					$scope,
-					$stmt,
-					'static',
-					'static variable',
-					true,
-				),
-			];
-
-			$vars = [];
-			foreach ($stmt->vars as $var) {
-				if (!is_string($var->var->name)) {
-					throw new ShouldNotHappenException();
-				}
-
-				if ($var->default !== null) {
-					$defaultExprResult = $this->processExprNode($stmt, $var->default, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-					$impurePoints = array_merge($impurePoints, $defaultExprResult->getImpurePoints());
-				}
-
-				$scope = $scope->enterExpressionAssign($var->var);
-				$varResult = $this->processExprNode($stmt, $var->var, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$impurePoints = array_merge($impurePoints, $varResult->getImpurePoints());
-				$scope = $scope->exitExpressionAssign($var->var);
-
-				$scope = $scope->assignVariable($var->var->name, new MixedType(), new MixedType(), TrinaryLogic::createYes());
-				$vars[] = $var->var->name;
-			}
-
-			$scope = $this->processVarAnnotation($scope, $vars, $stmt);
-		} elseif ($stmt instanceof Node\Stmt\Const_) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			foreach ($stmt->consts as $const) {
-				$this->callNodeCallback($nodeCallback, $const, $scope, $storage);
-				$constResult = $this->processExprNode($stmt, $const->value, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$impurePoints = array_merge($impurePoints, $constResult->getImpurePoints());
-				if ($const->namespacedName !== null) {
-					$constantName = new Name\FullyQualified($const->namespacedName->toString());
-				} else {
-					$constantName = new Name\FullyQualified($const->name->toString());
-				}
-				$scope = $scope->assignExpression(new ConstFetch($constantName), $constResult->getType(), $constResult->getNativeType());
-			}
-		} elseif ($stmt instanceof Node\Stmt\ClassConst) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [];
-			$this->processAttributeGroups($stmt, $stmt->attrGroups, $scope, $storage, $nodeCallback);
-			foreach ($stmt->consts as $const) {
-				$this->callNodeCallback($nodeCallback, $const, $scope, $storage);
-				$constResult = $this->processExprNode($stmt, $const->value, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$impurePoints = array_merge($impurePoints, $constResult->getImpurePoints());
-				if ($scope->getClassReflection() === null) {
-					throw new ShouldNotHappenException();
-				}
-				$scope = $scope->assignExpression(
-					new Expr\ClassConstFetch(new Name\FullyQualified($scope->getClassReflection()->getName()), $const->name),
-					$constResult->getType(),
-					$constResult->getNativeType(),
-				);
-			}
-		} elseif ($stmt instanceof Node\Stmt\EnumCase) {
-			$hasYield = false;
-			$throwPoints = [];
-			$this->processAttributeGroups($stmt, $stmt->attrGroups, $scope, $storage, $nodeCallback);
-			$impurePoints = [];
-			if ($stmt->expr !== null) {
-				$exprResult = $this->processExprNode($stmt, $stmt->expr, $scope, $storage, $nodeCallback, ExpressionContext::createDeep());
-				$impurePoints = $exprResult->getImpurePoints();
-			}
-		} elseif ($stmt instanceof InlineHTML) {
-			$hasYield = false;
-			$throwPoints = [];
-			$impurePoints = [
-				new ImpurePoint($scope, $stmt, 'betweenPhpTags', 'output between PHP opening and closing tags', true),
-			];
-		} elseif ($stmt instanceof Node\Stmt\Block) {
-			$result = $this->processStmtNodesInternal($stmt, $stmt->stmts, $scope, $storage, $nodeCallback, $context);
-			if ($this->polluteScopeWithBlock) {
-				return $result;
-			}
-
-			return new InternalStatementResult(
-				$scope->mergeWith($result->getScope()),
-				$result->hasYield(),
-				$result->isAlwaysTerminating(),
-				$result->getExitPoints(),
-				$result->getThrowPoints(),
-				$result->getImpurePoints(),
-				$result->getEndStatements(),
-			);
-		} elseif ($stmt instanceof Node\Stmt\Nop) {
-			$hasYield = false;
-			$throwPoints = $overridingThrowPoints ?? [];
-			$impurePoints = [];
-		} elseif ($stmt instanceof Node\Stmt\GroupUse) {
-			$hasYield = false;
-			$throwPoints = [];
-			foreach ($stmt->uses as $use) {
-				$this->callNodeCallback($nodeCallback, $use, $scope, $storage);
-			}
-			$impurePoints = [];
-		} else {
-			$hasYield = false;
-			$throwPoints = $overridingThrowPoints ?? [];
-			$impurePoints = [];
+			return $stmtResult;
 		}
 
-		return new InternalStatementResult($scope, $hasYield, false, [], $throwPoints, $impurePoints);
+		// statements with no analysis of their own (e.g. HaltCompiler)
+		return new InternalStatementResult($scope, hasYield: false, isAlwaysTerminating: false, exitPoints: [], throwPoints: $overridingThrowPoints ?? [], impurePoints: []);
 	}
 
 	/**
 	 * @return array{bool, string|null}
 	 */
-	private function getDeprecatedAttribute(Scope $scope, Node\Stmt\Function_|Node\Stmt\ClassMethod|Node\PropertyHook $stmt): array
+	public function getDeprecatedAttribute(Scope $scope, Node\Stmt\Function_|Node\Stmt\ClassMethod|Node\PropertyHook $stmt): array
 	{
 		$initializerExprContext = InitializerExprContext::fromStubParameter(
 			$scope->isInClass() ? $scope->getClassReflection()->getName() : null,
@@ -2767,48 +832,35 @@ class NodeScopeResolver
 		return null;
 	}
 
-	private function getCurrentClassReflection(Node\Stmt\ClassLike $stmt, string $className, Scope $scope): ClassReflection
+	public function isAnalysedFile(string $fileName): bool
 	{
-		if (!$this->reflectionProvider->hasClass($className)) {
-			return $this->createAstClassReflection($stmt, $className, $scope);
-		}
-
-		$defaultClassReflection = $this->reflectionProvider->getClass($className);
-		if ($defaultClassReflection->getFileName() !== $scope->getFile()) {
-			return $this->createAstClassReflection($stmt, $className, $scope);
-		}
-
-		$startLine = $defaultClassReflection->getNativeReflection()->getStartLine();
-		if ($startLine !== $stmt->getStartLine()) {
-			return $this->createAstClassReflection($stmt, $className, $scope);
-		}
-
-		return $defaultClassReflection;
+		return isset($this->analysedFiles[$fileName]);
 	}
 
-	private function createAstClassReflection(Node\Stmt\ClassLike $stmt, string $className, Scope $scope): ClassReflection
+	public function shouldPolluteScopeWithLoopInitialAssignments(): bool
 	{
-		$nodeToReflection = new NodeToReflection();
-		$betterReflectionClass = $nodeToReflection->__invoke(
-			$this->reflector,
-			$stmt,
-			new LocatedSource(FileReader::read($scope->getFile()), $className, $scope->getFile()),
-			$scope->getNamespace() !== null ? new Node\Stmt\Namespace_(new Name($scope->getNamespace())) : null,
-		);
-		if (!$betterReflectionClass instanceof \PHPStan\BetterReflection\Reflection\ReflectionClass) {
-			throw new ShouldNotHappenException();
-		}
+		return $this->polluteScopeWithLoopInitialAssignments;
+	}
 
-		return $this->classReflectionFactory->create(
-			$betterReflectionClass->getName(),
-			$betterReflectionClass instanceof ReflectionEnum && PHP_VERSION_ID >= 80000
-				? new \PHPStan\BetterReflection\Reflection\Adapter\ReflectionEnum($betterReflectionClass)
-				: new ReflectionClass($betterReflectionClass),
-			null,
-			null,
-			null,
-			sprintf('%s:%d', $scope->getFile(), $stmt->getStartLine()),
-		);
+	public function shouldPolluteScopeWithAlwaysIterableForeach(): bool
+	{
+		return $this->polluteScopeWithAlwaysIterableForeach;
+	}
+
+	public function shouldTreatPhpDocTypesAsCertain(): bool
+	{
+		return $this->treatPhpDocTypesAsCertain;
+	}
+
+	/** @return ExtensionsCollection<ReadWritePropertiesExtension> */
+	public function getReadWritePropertiesExtensions(): ExtensionsCollection
+	{
+		return $this->readWritePropertiesExtensions;
+	}
+
+	public function clearCalledMethodResults(): void
+	{
+		$this->calledMethodResults = [];
 	}
 
 	public function lookForSetAllowedUndefinedExpressions(MutatingScope $scope, Expr $expr): MutatingScope
@@ -3522,7 +1574,7 @@ class NodeScopeResolver
 	/**
 	 * @param callable(Node $node, Scope $scope): void $nodeCallback
 	 */
-	private function processParamNode(
+	public function processParamNode(
 		Node\Stmt $stmt,
 		Node\Param $param,
 		MutatingScope $scope,
@@ -3546,7 +1598,7 @@ class NodeScopeResolver
 	 * @param AttributeGroup[] $attrGroups
 	 * @param callable(Node $node, Scope $scope): void $nodeCallback
 	 */
-	private function processAttributeGroups(
+	public function processAttributeGroups(
 		Node\Stmt $stmt,
 		array $attrGroups,
 		MutatingScope $scope,
@@ -3588,7 +1640,7 @@ class NodeScopeResolver
 	 * @param Node\PropertyHook[] $hooks
 	 * @param callable(Node $node, Scope $scope): void $nodeCallback
 	 */
-	private function processPropertyHooks(
+	public function processPropertyHooks(
 		Node\Stmt $stmt,
 		Identifier|Name|ComplexType|null $nativeTypeNode,
 		?Type $phpDocType,
@@ -4737,605 +2789,6 @@ class NodeScopeResolver
 		return $scope;
 	}
 
-	/**
-	 * @return array{bodyScope: MutatingScope, endScope: MutatingScope, totalKeys: int}|null
-	 */
-	private function tryProcessUnrolledConstantArrayForeach(
-		Foreach_ $stmt,
-		MutatingScope $originalScope,
-		ExpressionResultStorage $originalStorage,
-		StatementContext $context,
-		Type $iterateeType,
-		Type $nativeIterateeType,
-	): ?array
-	{
-		if ($stmt->byRef) {
-			return null;
-		}
-		if (!($stmt->valueVar instanceof Variable && is_string($stmt->valueVar->name))) {
-			return null;
-		}
-		if ($stmt->keyVar !== null && !($stmt->keyVar instanceof Variable && is_string($stmt->keyVar->name))) {
-			return null;
-		}
-
-		if (!$iterateeType->isConstantArray()->yes()) {
-			return null;
-		}
-		$constantArrays = $iterateeType->getConstantArrays();
-		if (count($constantArrays) === 0) {
-			return null;
-		}
-
-		$totalKeys = 0;
-		$hasUnsealed = false;
-		foreach ($constantArrays as $constantArray) {
-			$totalKeys += count($constantArray->getKeyTypes());
-			if (!$constantArray->isUnsealed()->yes()) {
-				continue;
-			}
-			$hasUnsealed = true;
-		}
-		if ($totalKeys === 0 || $totalKeys > self::FOREACH_UNROLL_LIMIT) {
-			return null;
-		}
-		$foreachUnrollFactor = $context->getForeachUnrollFactor();
-		if ($foreachUnrollFactor > 1 && $foreachUnrollFactor * $totalKeys > self::FOREACH_UNROLL_NESTED_LIMIT) {
-			return null;
-		}
-
-		$nativeConstantArrays = $nativeIterateeType->getConstantArrays();
-		$matchedNativeArrays = count($nativeConstantArrays) === count($constantArrays) ? $nativeConstantArrays : null;
-
-		$valueVarName = $stmt->valueVar->name;
-		$keyVarName = $stmt->keyVar instanceof Variable ? $stmt->keyVar->name : null;
-
-		$allBodyScopes = [];
-		$allChainScopes = [];
-		$allBreakScopes = [];
-
-		$bodyContext = $context->enterUnrolledForeach($totalKeys);
-
-		foreach ($constantArrays as $arrayIndex => $constantArray) {
-			$keyTypes = $constantArray->getKeyTypes();
-			$valueTypes = $constantArray->getValueTypes();
-			if (count($keyTypes) === 0) {
-				continue;
-			}
-
-			$nativeConstantArray = $matchedNativeArrays !== null ? $matchedNativeArrays[$arrayIndex] : null;
-			$optionalKeys = array_fill_keys($constantArray->getOptionalKeys(), true);
-
-			$chainScope = $originalScope;
-			$entryScopes = [];
-
-			foreach ($keyTypes as $i => $keyType) {
-				$valueType = $valueTypes[$i];
-				$isOptional = isset($optionalKeys[$i]);
-
-				$nativeKeyType = $nativeConstantArray !== null && isset($nativeConstantArray->getKeyTypes()[$i])
-					? $nativeConstantArray->getKeyTypes()[$i]
-					: $keyType;
-				$nativeValueType = $nativeConstantArray !== null && isset($nativeConstantArray->getValueTypes()[$i])
-					? $nativeConstantArray->getValueTypes()[$i]
-					: $valueType;
-
-				$iterScope = $chainScope->assignVariable(
-					$valueVarName,
-					$valueType,
-					$nativeValueType,
-					TrinaryLogic::createYes(),
-				);
-				$iterScope = $iterScope->assignExpression(
-					new OriginalForeachValueExpr($valueVarName),
-					$valueType,
-					$nativeValueType,
-				);
-				if ($keyVarName !== null) {
-					$iterScope = $iterScope->assignVariable(
-						$keyVarName,
-						$keyType,
-						$nativeKeyType,
-						TrinaryLogic::createYes(),
-					);
-					$iterScope = $iterScope->assignExpression(
-						new OriginalForeachKeyExpr($keyVarName),
-						$keyType,
-						$nativeKeyType,
-					);
-					$iterScope = $iterScope->assignExpression(
-						new ArrayDimFetch($stmt->expr, $stmt->keyVar),
-						$valueType,
-						$nativeValueType,
-					);
-				}
-
-				$entryScopes[] = $iterScope;
-
-				$iterStorage = $originalStorage->duplicate();
-				$bodyResult = $this->processStmtNodesInternal(
-					$stmt,
-					$stmt->stmts,
-					$iterScope,
-					$iterStorage,
-					new NoopNodeCallback(),
-					$bodyContext,
-				)->filterOutLoopExitPoints();
-
-				$iterEndScope = $bodyResult->getScope();
-				foreach ($bodyResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-					$iterEndScope = $iterEndScope->mergeWith($continueExitPoint->getScope());
-				}
-				foreach ($bodyResult->getExitPointsByType(Break_::class) as $breakExitPoint) {
-					$allBreakScopes[] = $breakExitPoint->getScope();
-				}
-
-				if ($isOptional) {
-					$chainScope = $iterEndScope->mergeWith($chainScope);
-				} else {
-					$chainScope = $iterEndScope;
-				}
-			}
-
-			$arrayBodyScope = $entryScopes[0];
-			for ($i = 1, $c = count($entryScopes); $i < $c; $i++) {
-				$arrayBodyScope = $arrayBodyScope->mergeWith($entryScopes[$i]);
-			}
-			if (count($entryScopes) === 1) {
-				$arrayBodyScope = $arrayBodyScope->mergeWith($chainScope);
-			}
-
-			$allBodyScopes[] = $arrayBodyScope;
-			$allChainScopes[] = $chainScope;
-		}
-
-		if ($allBodyScopes === []) {
-			return null;
-		}
-
-		$bodyScope = $allBodyScopes[0];
-		for ($i = 1, $c = count($allBodyScopes); $i < $c; $i++) {
-			$bodyScope = $bodyScope->mergeWith($allBodyScopes[$i]);
-		}
-
-		$endScope = $allChainScopes[0];
-		for ($i = 1, $c = count($allChainScopes); $i < $c; $i++) {
-			$endScope = $endScope->mergeWith($allChainScopes[$i]);
-		}
-
-		foreach ($allBreakScopes as $breakScope) {
-			$endScope = $endScope->mergeWith($breakScope);
-		}
-
-		// Unsealed shapes describe zero-or-more additional entries beyond the
-		// explicit keys. Run the scope-generalizing loop on top of the
-		// unrolled explicit iterations so body-scope variables (e.g. counters)
-		// account for the extra iterations while keeping the lower bound
-		// established by the non-optional explicit keys.
-		if ($hasUnsealed) {
-			$loopScope = $endScope;
-			$count = 0;
-			do {
-				$prevLoopScope = $loopScope;
-				$iterStorage = $originalStorage->duplicate();
-				$iterBodyScope = $loopScope->mergeWith($endScope);
-				$iterBodyScope = $this->enterForeach($iterBodyScope, $iterStorage, $originalScope, $stmt, $iterateeType, $nativeIterateeType, new NoopNodeCallback());
-				$iterBodyScopeResult = $this->processStmtNodesInternal($stmt, $stmt->stmts, $iterBodyScope, $iterStorage, new NoopNodeCallback(), $context->enterDeep())->filterOutLoopExitPoints();
-				$loopScope = $iterBodyScopeResult->getScope();
-				foreach ($iterBodyScopeResult->getExitPointsByType(Continue_::class) as $continueExitPoint) {
-					$loopScope = $loopScope->mergeWith($continueExitPoint->getScope());
-				}
-				foreach ($iterBodyScopeResult->getExitPointsByType(Break_::class) as $breakExitPoint) {
-					$endScope = $endScope->mergeWith($breakExitPoint->getScope());
-				}
-				$bodyScope = $bodyScope->mergeWith($loopScope);
-				if ($loopScope->equals($prevLoopScope)) {
-					break;
-				}
-				if ($count >= self::GENERALIZE_AFTER_ITERATION) {
-					$loopScope = $prevLoopScope->generalizeWith($loopScope);
-				}
-				$count++;
-			} while ($count < self::LOOP_SCOPE_ITERATIONS);
-
-			$endScope = $endScope->mergeWith($loopScope);
-		}
-
-		return ['bodyScope' => $bodyScope, 'endScope' => $endScope, 'totalKeys' => $totalKeys];
-	}
-
-	private function getTraversableForeachThrowPoint(MutatingScope $scope, Expr $iteratee): ?InternalThrowPoint
-	{
-		$exprType = $scope->getType($iteratee);
-		$traversableType = new ObjectType(Traversable::class);
-
-		if ($traversableType->isSuperTypeOf($exprType)->no()) {
-			return null;
-		}
-
-		$traversablePart = TypeCombinator::intersect($exprType, $traversableType);
-		$iteratorAggregateType = new ObjectType(IteratorAggregate::class);
-
-		if ($iteratorAggregateType->isSuperTypeOf($traversablePart)->yes()
-			&& $traversablePart->hasMethod('getIterator')->yes()) {
-			$method = $traversablePart->getMethod('getIterator', $scope);
-			$throwType = $method->getThrowType();
-			if ($throwType !== null) {
-				if ($throwType->isVoid()->yes()) {
-					return null;
-				}
-				return InternalThrowPoint::createExplicit($scope, $throwType, $iteratee, true);
-			}
-
-			if (!$this->implicitThrows) {
-				return null;
-			}
-		}
-
-		return InternalThrowPoint::createImplicit($scope, $iteratee);
-	}
-
-	/**
-	 * @param callable(Node $node, Scope $scope): void $nodeCallback
-	 */
-	private function enterForeach(MutatingScope $scope, ExpressionResultStorage $storage, MutatingScope $originalScope, Foreach_ $stmt, Type $iterateeType, Type $nativeIterateeType, callable $nodeCallback): MutatingScope
-	{
-		if ($stmt->expr instanceof Variable && is_string($stmt->expr->name)) {
-			$scope = $this->processVarAnnotation($scope, [$stmt->expr->name], $stmt);
-		}
-
-		if (
-			($stmt->valueVar instanceof Variable && is_string($stmt->valueVar->name))
-			&& ($stmt->keyVar === null || ($stmt->keyVar instanceof Variable && is_string($stmt->keyVar->name)))
-		) {
-			$keyVarName = $stmt->keyVar instanceof Variable ? $stmt->keyVar->name : null;
-			$scope = $scope->enterForeach(
-				$originalScope,
-				$stmt->expr,
-				$iterateeType,
-				$nativeIterateeType,
-				$stmt->valueVar->name,
-				$keyVarName,
-				$stmt->byRef,
-			);
-			$vars = [$stmt->valueVar->name];
-			if ($keyVarName !== null) {
-				$vars[] = $keyVarName;
-			}
-		} else {
-			$scope = $this->processVirtualAssign(
-				$scope,
-				$storage,
-				$stmt,
-				$stmt->valueVar,
-				new NativeTypeExpr(
-					$originalScope->getIterableValueType($iterateeType),
-					$originalScope->getIterableValueType($nativeIterateeType),
-				),
-				$nodeCallback,
-			)->getScope();
-			$vars = $this->getAssignedVariables($stmt->valueVar);
-			if (
-				$stmt->keyVar instanceof Variable && is_string($stmt->keyVar->name)
-			) {
-				$scope = $scope->enterForeachKey($originalScope, $stmt->expr, $iterateeType, $nativeIterateeType, $stmt->keyVar->name);
-				$vars[] = $stmt->keyVar->name;
-			} elseif ($stmt->keyVar !== null) {
-				$scope = $this->processVirtualAssign(
-					$scope,
-					$storage,
-					$stmt,
-					$stmt->keyVar,
-					new NativeTypeExpr(
-						$originalScope->getIterableKeyType($iterateeType),
-						$originalScope->getIterableKeyType($nativeIterateeType),
-					),
-					$nodeCallback,
-				)->getScope();
-				$vars = array_merge($vars, $this->getAssignedVariables($stmt->keyVar));
-			}
-
-			if ($stmt->valueVar instanceof List_) {
-				$scope = $this->addDestructureTaggedUnionConditionalHolders(
-					$scope,
-					$originalScope->getIterableValueType($iterateeType),
-					$stmt->valueVar,
-				);
-			}
-		}
-
-		$constantArrays = $iterateeType->getConstantArrays();
-		if (
-			$stmt->getDocComment() === null
-			&& $iterateeType->isConstantArray()->yes()
-			&& count($constantArrays) === 1
-			&& $stmt->valueVar instanceof Variable && is_string($stmt->valueVar->name)
-			&& $stmt->keyVar instanceof Variable && is_string($stmt->keyVar->name)
-		) {
-			$valueConditionalHolders = [];
-			$arrayDimFetchConditionalHolders = [];
-			foreach ($constantArrays[0]->getKeyTypes() as $i => $keyType) {
-				$valueType = $constantArrays[0]->getValueTypes()[$i];
-				$keyExpressionTypeHolder = ExpressionTypeHolder::createYes(new Variable($stmt->keyVar->name), $keyType);
-
-				$holder = new ConditionalExpressionHolder([
-					'$' . $stmt->keyVar->name => $keyExpressionTypeHolder,
-				], ExpressionTypeHolder::createYes($stmt->valueVar, $valueType));
-				$valueConditionalHolders[$holder->getKey()] = $holder;
-				$arrayDimFetchHolder = new ConditionalExpressionHolder([
-					'$' . $stmt->keyVar->name => $keyExpressionTypeHolder,
-				], ExpressionTypeHolder::createYes(new ArrayDimFetch($stmt->expr, $stmt->keyVar), $valueType));
-				$arrayDimFetchConditionalHolders[$arrayDimFetchHolder->getKey()] = $arrayDimFetchHolder;
-			}
-
-			$scope = $scope->addConditionalExpressions(
-				'$' . $stmt->valueVar->name,
-				$valueConditionalHolders,
-			);
-			if ($stmt->expr instanceof Variable && is_string($stmt->expr->name)) {
-				$scope = $scope->addConditionalExpressions(
-					sprintf('$%s[$%s]', $stmt->expr->name, $stmt->keyVar->name),
-					$arrayDimFetchConditionalHolders,
-				);
-			}
-		}
-
-		if (
-			$stmt->expr instanceof FuncCall
-			&& $stmt->expr->name instanceof Name
-			&& !$stmt->expr->isFirstClassCallable()
-			&& $stmt->expr->name->toLowerString() === 'array_keys'
-			&& $stmt->valueVar instanceof Variable
-		) {
-			$args = $stmt->expr->getArgs();
-			if (count($args) >= 1) {
-				$arrayArg = $args[0]->value;
-				$scope = $scope->assignExpression(
-					new ArrayDimFetch($arrayArg, $stmt->valueVar),
-					$scope->getType($arrayArg)->getIterableValueType(),
-					$scope->getNativeType($arrayArg)->getIterableValueType(),
-				);
-			}
-		}
-
-		return $this->processVarAnnotation($scope, $vars, $stmt);
-	}
-
-	/**
-	 * When destructuring an iterable whose value type is a tagged union of
-	 * constant arrays — e.g. `array<array{null, int}|array{int, null}>` — the
-	 * variants describe a relationship between the destructured variables that
-	 * a per-variable narrowing would normally lose: knowing `$x === null` should
-	 * imply `$y === int`, but `foreach ($a as [$x, $y])` assigns `$x` and `$y`
-	 * independently, so each ends up as the union (`int|null`) and the link is
-	 * dropped.
-	 *
-	 * Recover the link by storing conditional-expression holders on each
-	 * destructured variable: for every variant, "when this variable matches the
-	 * variant's value at its position, the other variables match the variant's
-	 * values at their positions". A later `if ($x === null)` then fires the
-	 * matching holder and narrows `$y` accordingly.
-	 *
-	 * Only handles flat positional / keyed destructure patterns (List_) where
-	 * each item's target is a plain Variable; nested destructure is left for
-	 * the regular per-variable type tracking.
-	 */
-	private function addDestructureTaggedUnionConditionalHolders(
-		MutatingScope $scope,
-		Type $iterableValueType,
-		List_ $list,
-	): MutatingScope
-	{
-		$constantArrays = $iterableValueType->getConstantArrays();
-		if (count($constantArrays) < 2) {
-			return $scope;
-		}
-
-		// Collect each list item's array-key value and target variable.
-		$items = [];
-		foreach ($list->items as $position => $item) {
-			if ($item === null) {
-				continue;
-			}
-			if (!$item->value instanceof Variable || !is_string($item->value->name)) {
-				return $scope;
-			}
-			if ($item->key === null) {
-				$keyValue = $position;
-			} elseif ($item->key instanceof Node\Scalar\String_) {
-				$keyValue = $item->key->value;
-			} elseif ($item->key instanceof Node\Scalar\Int_) {
-				$keyValue = $item->key->value;
-			} else {
-				return $scope;
-			}
-			$items[] = ['key' => $keyValue, 'name' => $item->value->name];
-		}
-
-		if (count($items) < 2) {
-			return $scope;
-		}
-
-		// For every variant, every item must have a matching key with a single
-		// value type at it; otherwise the variants don't all describe the same
-		// destructure shape and we can't form a sound holder set.
-		$variantValuesByItem = [];
-		foreach ($items as $itemIdx => $itemInfo) {
-			$variantValuesByItem[$itemIdx] = [];
-			foreach ($constantArrays as $variantIdx => $variant) {
-				$keyType = is_int($itemInfo['key']) ? new ConstantIntegerType($itemInfo['key']) : new ConstantStringType($itemInfo['key']);
-				if (!$variant->hasOffsetValueType($keyType)->yes()) {
-					return $scope;
-				}
-				$variantValuesByItem[$itemIdx][$variantIdx] = $variant->getOffsetValueType($keyType);
-			}
-		}
-
-		// For each item × variant, build a holder: "when item is variant's value
-		// at this position, the *other* items are the variant's values at their
-		// positions". Skip the variant if the condition value is too wide to be
-		// a useful discriminator (i.e. equal to the union of all the variant
-		// values at this position — narrowing it back wouldn't pick a variant).
-		foreach ($items as $itemIdx => $itemInfo) {
-			$exprString = '$' . $itemInfo['name'];
-			$variantConditionTypes = $variantValuesByItem[$itemIdx];
-			$itemUnionType = TypeCombinator::union(...array_values($variantConditionTypes));
-			$holders = [];
-			foreach (array_keys($constantArrays) as $variantIdx) {
-				$conditionType = $variantConditionTypes[$variantIdx];
-				if ($conditionType->equals($itemUnionType)) {
-					continue;
-				}
-				$conditions = [
-					$exprString => ExpressionTypeHolder::createYes(new Variable($itemInfo['name']), $conditionType),
-				];
-				foreach ($items as $otherIdx => $otherInfo) {
-					if ($otherIdx === $itemIdx) {
-						continue;
-					}
-					$otherType = $variantValuesByItem[$otherIdx][$variantIdx];
-					$holder = new ConditionalExpressionHolder(
-						$conditions,
-						ExpressionTypeHolder::createYes(new Variable($otherInfo['name']), $otherType),
-					);
-					$holders['$' . $otherInfo['name']][$holder->getKey()] = $holder;
-				}
-			}
-
-			foreach ($holders as $targetExprString => $targetHolders) {
-				$scope = $scope->addConditionalExpressions($targetExprString, $targetHolders);
-			}
-		}
-
-		return $scope;
-	}
-
-	/**
-	 * @param callable(Node $node, Scope $scope): void $nodeCallback
-	 */
-	private function processTraitUse(Node\Stmt\TraitUse $node, MutatingScope $classScope, ExpressionResultStorage $storage, callable $nodeCallback): void
-	{
-		$parentTraitNames = [];
-		$parent = $classScope->getParentScope();
-		while ($parent !== null) {
-			if ($parent->isInTrait()) {
-				$parentTraitNames[] = $parent->getTraitReflection()->getName();
-			}
-			$parent = $parent->getParentScope();
-		}
-
-		foreach ($node->traits as $trait) {
-			$traitName = (string) $trait;
-			if (in_array($traitName, $parentTraitNames, true)) {
-				continue;
-			}
-			if (!$this->reflectionProvider->hasClass($traitName)) {
-				continue;
-			}
-			$traitReflection = $this->reflectionProvider->getClass($traitName);
-			$traitFileName = $traitReflection->getFileName();
-			if ($traitFileName === null) {
-				continue; // trait from eval or from PHP itself
-			}
-			$fileName = $this->fileHelper->normalizePath($traitFileName);
-			if (!isset($this->analysedFiles[$fileName])) {
-				continue;
-			}
-			$adaptations = [];
-			foreach ($node->adaptations as $adaptation) {
-				if ($adaptation->trait === null) {
-					$adaptations[] = $adaptation;
-					continue;
-				}
-				if ($adaptation->trait->toLowerString() !== $trait->toLowerString()) {
-					continue;
-				}
-
-				$adaptations[] = $adaptation;
-			}
-			$parserNodes = $this->parser->parseFile($fileName);
-			$this->processNodesForTraitUse($parserNodes, $traitReflection, $classScope, $storage, $adaptations, $nodeCallback);
-		}
-	}
-
-	/**
-	 * @param Node[]|Node|scalar|null $node
-	 * @param Node\Stmt\TraitUseAdaptation[] $adaptations
-	 * @param callable(Node $node, Scope $scope): void $nodeCallback
-	 */
-	private function processNodesForTraitUse($node, ClassReflection $traitReflection, MutatingScope $scope, ExpressionResultStorage $storage, array $adaptations, callable $nodeCallback): void
-	{
-		if ($node instanceof Node) {
-			if ($node instanceof Node\Stmt\Trait_ && $traitReflection->getName() === (string) $node->namespacedName && $traitReflection->getNativeReflection()->getStartLine() === $node->getStartLine()) {
-				$methodModifiers = [];
-				$methodNames = [];
-				foreach ($adaptations as $adaptation) {
-					if (!$adaptation instanceof Node\Stmt\TraitUseAdaptation\Alias) {
-						continue;
-					}
-
-					$methodName = $adaptation->method->toLowerString();
-					if ($adaptation->newModifier !== null) {
-						$methodModifiers[$methodName] = $adaptation->newModifier;
-					}
-
-					if ($adaptation->newName === null) {
-						continue;
-					}
-
-					$methodNames[$methodName] = $adaptation->newName;
-				}
-
-				$stmts = $node->stmts;
-				foreach ($stmts as $i => $stmt) {
-					if (!$stmt instanceof Node\Stmt\ClassMethod) {
-						continue;
-					}
-					$methodName = $stmt->name->toLowerString();
-					$methodAst = clone $stmt;
-					$stmts[$i] = $methodAst;
-					if (array_key_exists($methodName, $methodModifiers)) {
-						$methodAst->flags = ($methodAst->flags & ~ Modifiers::VISIBILITY_MASK) | $methodModifiers[$methodName];
-					}
-
-					if (!array_key_exists($methodName, $methodNames)) {
-						continue;
-					}
-
-					$methodAst->setAttribute('originalTraitMethodName', $methodAst->name->toLowerString());
-					$methodAst->name = $methodNames[$methodName];
-				}
-
-				if (!$scope->isInClass()) {
-					throw new ShouldNotHappenException();
-				}
-				$traitScope = $scope->enterTrait($traitReflection);
-
-				// attribute args are not processed as part of the trait statements
-				// but rules like TraitAttributesRule ask about their types
-				$this->processAttributeGroups($node, $node->attrGroups, $traitScope, $storage, new NoopNodeCallback());
-
-				$this->callNodeCallback($nodeCallback, new InTraitNode($node, $traitReflection, $scope->getClassReflection()), $traitScope, $storage);
-				$this->processStmtNodesInternal($node, $stmts, $traitScope, $storage, $nodeCallback, StatementContext::createTopLevel());
-				return;
-			}
-			if ($node instanceof Node\Stmt\ClassLike) {
-				return;
-			}
-			if ($node instanceof Node\FunctionLike) {
-				return;
-			}
-			foreach ($node->getSubNodeNames() as $subNodeName) {
-				$subNode = $node->{$subNodeName};
-				$this->processNodesForTraitUse($subNode, $traitReflection, $scope, $storage, $adaptations, $nodeCallback);
-			}
-		} elseif (is_array($node)) {
-			foreach ($node as $subNode) {
-				$this->processNodesForTraitUse($subNode, $traitReflection, $scope, $storage, $adaptations, $nodeCallback);
-			}
-		}
-	}
-
 	public function processCalledMethod(MethodReflection $methodReflection): ?MutatingScope
 	{
 		$declaringClass = $methodReflection->getDeclaringClass();
@@ -5745,88 +3198,6 @@ class NodeScopeResolver
 			$isPassedUnreachableStatement = true;
 		}
 		return $stmts;
-	}
-
-	private function inferForLoopExpressions(For_ $stmt, Expr $lastCondExpr, MutatingScope $bodyScope): MutatingScope
-	{
-		// infer $items[$i] type from for ($i = 0; $i < count($items); $i++) {...}
-
-		if (
-			// $i = 0
-			count($stmt->init) === 1
-			&& $stmt->init[0] instanceof Assign
-			&& $stmt->init[0]->var instanceof Variable
-			&& $stmt->init[0]->expr instanceof Node\Scalar\Int_
-			&& $stmt->init[0]->expr->value === 0
-			// $i++ or ++$i
-			&& count($stmt->loop) === 1
-			&& ($stmt->loop[0] instanceof Expr\PreInc || $stmt->loop[0] instanceof Expr\PostInc)
-			&& $stmt->loop[0]->var instanceof Variable
-		) {
-			// $i < count($items)
-			if (
-				$lastCondExpr instanceof BinaryOp\Smaller
-				&& $lastCondExpr->left instanceof Variable
-				&& $lastCondExpr->right instanceof FuncCall
-				&& $lastCondExpr->right->name instanceof Name
-				&& !$lastCondExpr->right->isFirstClassCallable()
-				&& in_array($lastCondExpr->right->name->toLowerString(), ['count', 'sizeof'], true)
-				&& count($lastCondExpr->right->getArgs()) > 0
-				&& $lastCondExpr->right->getArgs()[0]->value instanceof Variable
-				&& is_string($stmt->init[0]->var->name)
-				&& $stmt->init[0]->var->name === $stmt->loop[0]->var->name
-				&& $stmt->init[0]->var->name === $lastCondExpr->left->name
-			) {
-				$arrayArg = $lastCondExpr->right->getArgs()[0]->value;
-				$arrayType = $bodyScope->getType($arrayArg);
-				if ($arrayType->isList()->yes()) {
-					$bodyScope = $bodyScope->assignExpression(
-						new ArrayDimFetch($lastCondExpr->right->getArgs()[0]->value, $lastCondExpr->left),
-						$arrayType->getIterableValueType(),
-						$bodyScope->getNativeType($arrayArg)->getIterableValueType(),
-					);
-				}
-			}
-
-			// count($items) > $i
-			if (
-				$lastCondExpr instanceof BinaryOp\Greater
-				&& $lastCondExpr->right instanceof Variable
-				&& $lastCondExpr->left instanceof FuncCall
-				&& $lastCondExpr->left->name instanceof Name
-				&& !$lastCondExpr->left->isFirstClassCallable()
-				&& in_array($lastCondExpr->left->name->toLowerString(), ['count', 'sizeof'], true)
-				&& count($lastCondExpr->left->getArgs()) > 0
-				&& $lastCondExpr->left->getArgs()[0]->value instanceof Variable
-				&& is_string($stmt->init[0]->var->name)
-				&& $stmt->init[0]->var->name === $stmt->loop[0]->var->name
-				&& $stmt->init[0]->var->name === $lastCondExpr->right->name
-			) {
-				$arrayArg = $lastCondExpr->left->getArgs()[0]->value;
-				$arrayType = $bodyScope->getType($arrayArg);
-				if ($arrayType->isList()->yes()) {
-					$bodyScope = $bodyScope->assignExpression(
-						new ArrayDimFetch($lastCondExpr->left->getArgs()[0]->value, $lastCondExpr->right),
-						$arrayType->getIterableValueType(),
-						$bodyScope->getNativeType($arrayArg)->getIterableValueType(),
-					);
-				}
-			}
-		}
-
-		return $bodyScope;
-	}
-
-	private function getGlobalVariableType(string $variableName): Type
-	{
-		if ($variableName === 'argc') {
-			return StaticTypeFactory::argc();
-		}
-		if ($variableName === 'argv') {
-			return StaticTypeFactory::argv();
-		}
-
-		return new MixedType();
 	}
 
 }
