+++ src/Analyser/StmtHandler/GroupUseHandler.php
@@ -1,0 +1,44 @@
+<?php declare(strict_types = 1);
+
+namespace PHPStan\Analyser\StmtHandler;
+
+use PhpParser\Node\Stmt;
+use PhpParser\Node\Stmt\GroupUse;
+use PHPStan\Analyser\ExpressionResultStorage;
+use PHPStan\Analyser\InternalStatementResult;
+use PHPStan\Analyser\MutatingScope;
+use PHPStan\Analyser\NodeScopeResolver;
+use PHPStan\Analyser\StatementContext;
+use PHPStan\Analyser\StmtHandler;
+use PHPStan\DependencyInjection\AutowiredService;
+
+/**
+ * @implements StmtHandler<GroupUse>
+ */
+#[AutowiredService]
+final class GroupUseHandler implements StmtHandler
+{
+
+	public function supports(Stmt $stmt): bool
+	{
+		return $stmt instanceof GroupUse;
+	}
+
+	public function processStmt(
+		NodeScopeResolver $nodeScopeResolver,
+		Stmt $stmt,
+		MutatingScope $scope,
+		ExpressionResultStorage $storage,
+		callable $nodeCallback,
+		StatementContext $context,
+	): InternalStatementResult
+	{
+		foreach ($stmt->uses as $use) {
+			$nodeScopeResolver->callNodeCallback($nodeCallback, $use, $scope, $storage);
+		}
+		$impurePoints = [];
+
+		return new InternalStatementResult($scope, hasYield: false, isAlwaysTerminating: false, exitPoints: [], throwPoints: [], impurePoints: $impurePoints);
+	}
+
+}
