+++ tests/PHPStan/Analyser/Fiber/FiberNodeScopeResolverRuleTest.php
@@ -4,14 +4,12 @@ namespace PHPStan\Analyser\Fiber;
 
 use PhpParser\Node;
 use PHPStan\Analyser\ExpressionResultFactory;
-use PHPStan\Analyser\ExprHandler\Helper\ImplicitToStringCallHelper;
 use PHPStan\Analyser\NodeScopeResolver;
 use PHPStan\Analyser\PerFileAnalysisResettable;
 use PHPStan\Analyser\Scope;
 use PHPStan\DependencyInjection\DirectExtensionsCollection;
 use PHPStan\File\FileHelper;
 use PHPStan\PhpDoc\PhpDocInheritanceResolver;
-use PHPStan\Reflection\ClassReflectionFactory;
 use PHPStan\Reflection\InitializerExprTypeResolver;
 use PHPStan\Rules\IdentifierRuleError;
 use PHPStan\Rules\Properties\ReadWritePropertiesExtension;
@@ -124,8 +122,6 @@ class FiberNodeScopeResolverRuleTest extends RuleTestCase
 			self::getContainer(),
 			$reflectionProvider,
 			self::getContainer()->getByType(InitializerExprTypeResolver::class),
-			self::getReflector(),
-			self::getContainer()->getByType(ClassReflectionFactory::class),
 			self::getContainer()->getExtensionsCollection(FunctionParameterOutTypeExtension::class),
 			self::getContainer()->getExtensionsCollection(MethodParameterOutTypeExtension::class),
 			self::getContainer()->getExtensionsCollection(StaticMethodParameterOutTypeExtension::class),
@@ -133,7 +129,6 @@ class FiberNodeScopeResolverRuleTest extends RuleTestCase
 			self::getContainer()->getByType(FileTypeMapper::class),
 			self::getContainer()->getByType(PhpDocInheritanceResolver::class),
 			self::getContainer()->getByType(FileHelper::class),
-			$typeSpecifier,
 			$readWritePropertiesExtensions !== [] ? new DirectExtensionsCollection($readWritePropertiesExtensions) : self::getContainer()->getExtensionsCollection(ReadWritePropertiesExtension::class),
 			self::getContainer()->getExtensionsCollection(FunctionParameterClosureThisExtension::class),
 			self::getContainer()->getExtensionsCollection(MethodParameterClosureThisExtension::class),
@@ -145,10 +140,8 @@ class FiberNodeScopeResolverRuleTest extends RuleTestCase
 			self::createScopeFactory($reflectionProvider, $typeSpecifier),
 			$this->shouldPolluteScopeWithLoopInitialAssignments(),
 			$this->shouldPolluteScopeWithAlwaysIterableForeach(),
-			self::getContainer()->getParameter('polluteScopeWithBlock'),
 			self::getContainer()->getParameter('exceptions')['implicitThrows'],
 			$this->shouldTreatPhpDocTypesAsCertain(),
-			self::getContainer()->getByType(ImplicitToStringCallHelper::class),
 			self::getContainer()->getByType(ExpressionResultFactory::class),
 		);
 	}
